% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws}
\alias{as_draws}
\title{Transform to \code{draws} objects}
\usage{
as_draws(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Try to transform an \R object to a \code{draws} object
supported by the \pkg{posterior} package.
}
\details{
The \code{draws} class is the parent class of
all draws formats supported by the \pkg{posterior} package.
}
