% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{as_posterior}
\alias{as_posterior}
\title{Attempt to transform to a posterior object}
\usage{
as_posterior(x, class = NULL)
}
\arguments{
\item{x}{An \R object to be converted.}

\item{class}{optional posterior class to transform to.
If \code{NULL} (the default), the class most similar to the
structure of \code{x} is chosen.}
}
\description{
Attempt to transform to a posterior object
}
