% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_draws.R
\name{as_posterior_draws}
\alias{as_posterior_draws}
\title{Attempt to transform to a posterior_draws object}
\usage{
as_posterior_draws(x, ...)
}
\arguments{
\item{x}{An \R object to be transformed.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Attempt to transform to a posterior_draws object
}
