% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{autocovariance}
\alias{autocovariance}
\title{Autocovariance estimates}
\usage{
autocovariance(x)
}
\arguments{
\item{x}{A numeric vector forming a sequence of values.}
}
\value{
A numeric vector of autocovariances at every lag (scaled by N-lag).
}
\description{
Compute autocovariance estimates for every lag for the specified
input sequence using a fast Fourier transform approach. Estimate
for lag t is scaled by N-t.
}
