% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_draws.R
\name{bind_draws}
\alias{bind_draws}
\alias{bind_draws.draws_matrix}
\alias{bind_draws.draws_array}
\alias{bind_draws.draws_df}
\alias{bind_draws.draws_list}
\alias{bind_draws.draws_rvars}
\title{Bind \code{draws} objects together}
\usage{
bind_draws(x, ...)

\method{bind_draws}{draws_matrix}(x, ..., along = "variable")

\method{bind_draws}{draws_array}(x, ..., along = "variable")

\method{bind_draws}{draws_df}(x, ..., along = "variable")

\method{bind_draws}{draws_list}(x, ..., along = "variable")

\method{bind_draws}{draws_rvars}(x, ..., along = "variable")
}
\arguments{
\item{x}{A \code{\link{draws}} object. The draws format of \code{x} will define the format
of the returned draws object.}

\item{...}{Further \code{\link{draws}} objects.}

\item{along}{The dimenion along with draws objects should be bind together.
Possible values are \code{variable} (the default), \code{chain}, \code{iteration}, and
\code{draw}. Not all options are supported for all input formats.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Bind multiple \code{\link{draws}} objects together to form a single
\code{draws} object.
}
