% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{diagnostics}
\alias{diagnostics}
\alias{convergence}
\title{Convergence diagnostics}
\description{
A list of available diagnostics and links to their individual help pages.
}
\details{
\itemize{
\item \code{\link[=ess_basic]{ess_basic()}}: Basic version of effective sample size
\item \code{\link[=ess_bulk]{ess_bulk()}}: Bulk effective sample size
\item \code{\link[=ess_tail]{ess_tail()}}: Tail effective sample size
\item \code{\link[=ess_quantile]{ess_quantile()}}: Effective sample sizes for quantiles
\item \code{\link[=ess_sd]{ess_sd()}}: Effective sample sizes for standard deviations
\item \code{\link[=mcse_mean]{mcse_mean()}}: Monte Carlo standard error for the mean
\item \code{\link[=mcse_quantile]{mcse_quantile()}}: Monte Carlo standard error for quantiles
\item \code{\link[=mcse_sd]{mcse_sd()}}: Monte Carlo standard error for standard deviations
\item \code{\link[=rhat_basic]{rhat_basic()}}: Basic version of Rhat
\item \code{\link[=rhat]{rhat()}}: Improved, rank-based version of Rhat
\item \code{\link[=rstar]{rstar()}}: R* diagnostic
}
}
