% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{.rhat}
\alias{.rhat}
\title{Compute the Rhat convergence diagnostic}
\usage{
.rhat(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute the Rhat convergence diagnostic
}
\keyword{internal}
