% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws_array.R
\name{draws_array}
\alias{draws_array}
\alias{as_draws_array}
\alias{as_draws_array.default}
\alias{as_draws_array.draws_array}
\alias{as_draws_array.draws_matrix}
\alias{as_draws_array.draws_df}
\alias{as_draws_array.draws_list}
\alias{is_draws_array}
\title{The \code{draws_array} format}
\usage{
as_draws_array(x, ...)

\method{as_draws_array}{default}(x, ...)

\method{as_draws_array}{draws_array}(x, ...)

\method{as_draws_array}{draws_matrix}(x, ...)

\method{as_draws_array}{draws_df}(x, ...)

\method{as_draws_array}{draws_list}(x, ...)

is_draws_array(x)
}
\arguments{
\item{x}{An object to convert to a \code{draws_array} object.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws_array} object, which has classes
\code{c("draws_array", "draws", "array")}.
}
\description{
The \code{as_draws_array()} methods convert
objects to the \code{draws_array} format. See \strong{Details}.
}
\details{
Objects of class \code{"draws_array"} are 3-D arrays with dimensions
\code{"iteration"}, \code{"chain"}, and \code{"variable"}. See \strong{Examples}.
}
\examples{
x <- as_draws_array(example_draws())
class(x)
str(x)
}
\seealso{
Other formats: 
\code{\link{draws_df}},
\code{\link{draws_list}},
\code{\link{draws_matrix}},
\code{\link{draws}}
}
\concept{formats}
