% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws_df.R
\name{draws_df}
\alias{draws_df}
\alias{as_draws_df}
\alias{as_draws_df.default}
\alias{as_draws_df.data.frame}
\alias{as_draws_df.draws_df}
\alias{as_draws_df.draws_matrix}
\alias{as_draws_df.draws_array}
\alias{as_draws_df.draws_list}
\alias{as_draws_df.mcmc}
\alias{as_draws_df.mcmc.list}
\alias{is_draws_df}
\title{The \code{draws_df} format}
\usage{
as_draws_df(x, ...)

\method{as_draws_df}{default}(x, ...)

\method{as_draws_df}{data.frame}(x, .iteration = NULL, .chain = NULL, ...)

\method{as_draws_df}{draws_df}(x, ...)

\method{as_draws_df}{draws_matrix}(x, ...)

\method{as_draws_df}{draws_array}(x, ...)

\method{as_draws_df}{draws_list}(x, ...)

\method{as_draws_df}{mcmc}(x, ...)

\method{as_draws_df}{mcmc.list}(x, ...)

is_draws_df(x)
}
\arguments{
\item{x}{An object to convert to a \code{draws_df} object.}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{.iteration}{Optional name of a column in the supplied \code{data.frame}
that contain iteration indices. If \code{NULL} (the default), the \code{.iteration}
column is used if it exists. Otherwise, the input is treated as being
index continuously from the first to the last row.}

\item{.chain}{Optional name of a column in the supplied \code{data.frame}
that contain chain indices. If \code{NULL} (the default), the \code{.chain}
column is used if it exists. Otherwise, the input is treated as belonging
to a single chain.}
}
\value{
A \code{draws_df} object, which has classes
\code{c("draws_df", "draws", class(tibble::tibble()))}.
}
\description{
The \code{as_draws_df()} methods convert
objects to the \code{draws_df} format. See \strong{Details}.
}
\details{
Objects of class \code{"draws_df"} are \link[tibble:tibble]{tibble} data
frames. They have one column per variable as well as additional metadata
columns \code{".iteration"}, \code{".chain"}, and \code{".draw"}. The difference between
the \code{".iteration"} and \code{".draw"} columns is that the former is relative to
the MCMC chain while the latter ignores the chain information and has all
unique values. See \strong{Examples}.
}
\examples{
x <- as_draws_df(example_draws())
class(x)
str(x)

# the difference between iteration and draw is clearer when contrasting
# the head and tail of the data frame
head(x[, c(".iteration", ".chain", ".draw")])
tail(x[, c(".iteration", ".chain", ".draw")])

}
\seealso{
Other formats: 
\code{\link{draws_array}},
\code{\link{draws_list}},
\code{\link{draws_matrix}},
\code{\link{draws}}
}
\concept{formats}
