% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{draws_eight_schools}
\alias{draws_eight_schools}
\alias{eight_schools}
\title{Draws from the eight schools model}
\format{A \code{draws_array} object with 100 iterations for each of the 4
chains containing posterior draws of the following variables.
\describe{
\item{mu}{Overall mean of the eight schools}
\item{tau}{Standard deviation between schools}
\item{theta}{Individual means of each of the eight schools}
}}
\source{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
\usage{
draws_eight_schools
}
\description{
Posterior draws from the eight schools
model described in Gelman et al. 2013.
}
\examples{
data("draws_eight_schools")
summarise_draws(draws_eight_schools)

}
\keyword{datasets}
