% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws_list.R
\name{draws_list}
\alias{draws_list}
\alias{as_draws_list}
\alias{as_draws_list.default}
\alias{as_draws_list.draws_list}
\alias{as_draws_list.draws_matrix}
\alias{as_draws_list.draws_array}
\alias{as_draws_list.draws_df}
\title{The \code{draws_list} format}
\usage{
as_draws_list(x, ...)

\method{as_draws_list}{default}(x, ...)

\method{as_draws_list}{draws_list}(x, ...)

\method{as_draws_list}{draws_matrix}(x, ...)

\method{as_draws_list}{draws_array}(x, ...)

\method{as_draws_list}{draws_df}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \code{draws_list} object.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws_list} object, which has classes \code{c("draws_list", "draws", "list")}.
}
\description{
Convert objects to the \code{draws_list} format.
}
\details{
Objects of class \code{"draws_list"} are lists with one element per MCMC
chain. Each of these elements is itself a named list of numeric vectors
with one vector per variable. The length of each vector is equal to the
number of saved iterations per chain. See \strong{Examples}.
}
\examples{
x <- as_draws_list(draws_eight_schools)
class(x)
str(x)

}
