% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-.R
\name{draws_of}
\alias{draws_of}
\alias{draws_of<-}
\title{Get/set array of draws underlying a random variable}
\usage{
draws_of(x)

draws_of(x) <- value
}
\arguments{
\item{x}{An \link{rvar}}

\item{value}{An array}
}
\description{
Gets/sets the array-representation that backs an \link{rvar}
}
\details{
While \link{rvar}s implements fast versions of basic math operations (including
\link[=rvar-matmult]{matrix multiplication}), sometimes you may need to bypass
the \link{rvar} abstraction to do what you need to do more efficiently.
\code{draws_of()} allows you to get / set the underlying array of draws in
order to do that.

\link{rvar}s represent draws internally using arrays of arbitrary dimension, which
is returned by \code{draws_of(x)} and can be set using \code{draws_of(x) <- value}.
The \strong{last} dimension of these arrays is the index of the draws.
}
