% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws_rvar.R
\name{draws_rvar}
\alias{draws_rvar}
\alias{as_draws_rvar}
\alias{as_draws_rvar.default}
\alias{as_draws_rvar.draws_rvar}
\alias{as_draws_rvar.rvar}
\alias{is_draws_rvar}
\title{The \code{draws_rvar} format}
\usage{
as_draws_rvar(x, ...)

\method{as_draws_rvar}{default}(x, ...)

\method{as_draws_rvar}{draws_rvar}(x, ...)

\method{as_draws_rvar}{rvar}(x, ...)

draws_rvar(..., .nchains = 1)

is_draws_rvar(x)
}
\arguments{
\item{x}{An object to convert to a \code{draws_rvar} object.}

\item{...}{For \code{as_draws_rvar()}:
Arguments passed to individual methods (if applicable).
For \code{draws_rvar()}: Named
arguments containing numeric vectors each defining a separate variable.}

\item{.nchains}{The number of chains (default is \code{1}).}
}
\value{
A \code{draws_rvar} object, which has classes
\code{c("draws_rvar", "draws", "list")}.
}
\description{
The \code{as_draws_rvar()} methods convert
objects to the \code{draws_rvar} format.
The \code{draws_rvar()} function creates an object of the
\code{draws_rvar} format based on a set of numeric vectors.
See \strong{Details}.
}
\details{
Objects of class \code{"draws_rvar"} are lists of \code{\link{rvar}} objects.
See \strong{Examples}.
}
\examples{
x1 <- as_draws_rvar(example_draws())
class(x1)
print(x1)
str(x1)

x2 <- draws_rvar(a = rnorm(10), b = rnorm(10), c = 1)
class(x2)
print(x2)
str(x2)
}
\seealso{
Other formats: 
\code{\link{draws_array}()},
\code{\link{draws_df}()},
\code{\link{draws_list}()},
\code{\link{draws_matrix}()},
\code{\link{draws}}
}
\concept{formats}
