% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_draws.R
\name{draws_summary}
\alias{draws_summary}
\alias{summarise_draws}
\alias{summarize_draws}
\alias{summarise_draws.draws}
\alias{summary.draws}
\alias{default_summary_measures}
\alias{default_convergence_measures}
\alias{default_mcse_measures}
\title{Summaries of draws objects}
\usage{
summarise_draws(x, ...)

summarize_draws(x, ...)

\method{summarise_draws}{draws}(x, measures = NULL, ...,
  probs = c(0.05, 0.95))

\method{summary}{draws}(x, ...)

default_summary_measures()

default_convergence_measures()

default_mcse_measures()
}
\arguments{
\item{x}{A \code{draws} object or one coercible to a \code{draws} object.}

\item{...}{Optionally, arguments to pass to specific methods.}

\item{measures}{A character vector containing the names of summary stats or
diagnostics to include. The convenience functions with names \code{default_*}
return character vectors with the names of the measures included by
default.}

\item{probs}{A numeric vector of probabilities used to compute quantiles. In
the output any columns corresponding to quantiles will have names starting
with lowercase \code{"q"} (e.g. \code{"q95"}).}
}
\value{
The \code{summarise_draws()} methods return a \link[tibble:tibble]{tibble} data frame.
The first column, \code{"variable"}, contains the variable names and the remaining
columns contain summary statistics and diagnostics.

The functions \code{default_summary_measures()}, \code{default_convergence_measures()},
and \code{default_mcse_measures()} return character vectors of names of the
default measures included.
}
\description{
The \code{summarise_draws()} (and \code{summarize_draws()}) methods provide a quick way
to get a table of summary statistics and diagnostics. These methods will
convert an object to a \code{draws} object if it isn't already. For convenience, a
\link[base:summary]{summary()} method for \code{draws} objects is also provided as an
alias for \code{summarise_draws()} if the input object is already a \code{draws}
object.
}
\examples{
x <- example_draws()
class(x)
str(x)

summarise_draws(x)
summarise_draws(x, "quantile", probs = c(0.1, 0.9))
summarise_draws(x, c("mean", "median"))

}
