% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-math.R
\name{E}
\alias{E}
\alias{Pr}
\title{Expectations of random variables}
\usage{
E(x, na.rm = FALSE)

Pr(x, na.rm = FALSE)
}
\arguments{
\item{x}{an \code{\link[=rvar]{rvar()}}}

\item{na.rm}{Should \code{NA} values in the random variable be removed before
computing expectations?

Both \code{E()} and \code{Pr()} take means over the draws dimension of the provided
random variable. \code{Pr()} additionally checks that the provided \link{rvar}
is a logical variable (hence, taking its expectation results in a probability).}
}
\value{
A numeric vector with the same dimensions as the given random variable, where
each entry in the vector is the mean of the corresponding entry in \code{x}.
}
\description{
Compute expectations and probabilities of a random variable.
}
\examples{

x = rvar(rnorm(4000, mean = 1, sd = 2))

E(x) # should be ~= 1

Pr(x < 0.5)
# is equivalent to this:
cdf(x, 0.5)
# and should be about the same as this:
pnorm(0.5, mean = 1, sd = 2)

}
