% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-summaries.R
\name{entropy}
\alias{entropy}
\title{Normalized entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{(multiple options) A vector type to be interpreted as draws from
a categorical distribution, such as:
\itemize{
\item A \link{factor}
\item A \link{numeric} (should be \link{integer} or integer-like)
\item An \link{rvar}, \link{rvar_factor}, or \link{rvar_ordered}
}}
}
\value{
Shannon entropy of the draws in \code{x}, normalized to be between 0 (all probability
in one category) and 1 (uniform):

\deqn{-\frac{\sum_{i = 1}^{n} p_i \log(p_i)}{\log(n)}}

If \code{x} is an \link{rvar}, returns an array of the same shape as \code{x}, where each
cell is the normalized Shannon entropy of the corresponding cell in \code{x}.
}
\description{
Normalized entropy, for measuring dispersion in draws from categorical distributions.
}
