% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_basic}
\alias{ess_basic}
\title{Basic version of the effective sample size}
\usage{
ess_basic(x, split = TRUE)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}

\item{split}{Logical. If \code{TRUE}, compute ESS on split chains.}
}
\value{
A single numeric value for the effective sample size.
}
\description{
Compute the basic effective sample size (ESS) estimate for a single parameter
as described in Gelman et al. (2013). For practical applications, we strongly
recommend the improved ESS convergence diagnostics implemented in
\code{\link{ess_bulk}} and \code{\link{ess_tail}}.
}
\references{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
