% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_mean}
\alias{ess_mean}
\title{Effective sample size for the mean}
\usage{
ess_mean(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute effective sample size estimate for a mean (expectation)
estimate of a single variable.
}
\examples{
mu <- extract_one_variable_matrix(example_draws(), "mu")
ess_mean(mu)

}
\references{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
