% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_quantile}
\alias{ess_quantile}
\alias{ess_median}
\title{Quantile effective sample size}
\usage{
ess_quantile(x, probs, names = TRUE)

ess_median(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}

\item{probs}{A numeric vector of probabilities.}
}
\value{
A numeric vector of effective sample sizes for quantile estimates.
}
\description{
Compute effective sample size estimates for quantile estimates of
a single parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
