% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_rfun}
\alias{ess_rfun}
\title{Effective sample size}
\usage{
ess_rfun(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}
}
\value{
A single numeric value for the effective sample size.
}
\description{
Compute the effective sample size estimate for a sample of several chains
for one parameter. For split-ESS, call this with split chains.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
