% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{extract_one_par_matrix}
\alias{extract_one_par_matrix}
\title{Extract posterior matrix of a single parameter}
\usage{
extract_one_par_matrix(x, par)
}
\arguments{
\item{x}{A posterior object.}

\item{par}{A parameter name to extract draws for.}
}
\description{
extract a draws x chain matrix for a single parameter
required for various convergence diagnostics.
}
