% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{extract_one_variable_matrix}
\alias{extract_one_variable_matrix}
\title{Extract posterior matrix of a single parameter}
\usage{
extract_one_variable_matrix(x, variable, ...)
}
\arguments{
\item{x}{A posterior_draws object.}

\item{variable}{A parameter name to extract draws for.}
}
\description{
extract a iteration x chain matrix for a single parameter
required for various convergence diagnostics.
}
