% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variable.R
\name{extract_variable}
\alias{extract_variable}
\alias{extract_variable.default}
\alias{extract_variable.draws}
\title{Extract draws of a single variable}
\usage{
extract_variable(x, variable, ...)

\method{extract_variable}{default}(x, variable, ...)

\method{extract_variable}{draws}(x, variable, ...)
}
\arguments{
\item{x}{A \code{draws} object, or another \R object for which the method is defined.}

\item{variable}{Name of a single variable to extract draws for.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A numeric vector of length #draws.
}
\description{
Extract a vector of draws of a single variable.
}
\examples{
x <- example_draws()
mu <- extract_variable(x, variable = "mu")
str(mu)

}
