% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{fold_draws}
\alias{fold_draws}
\title{Fold draws around their median}
\usage{
fold_draws(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
An array or vector of folded draws.
}
\description{
Fold draws around their median
}
\keyword{internal}
