% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices.R
\name{get_indices}
\alias{get_indices}
\title{Parse the indices from a list of variables (extracted e.g. from a
\code{draws_summary} object or a \code{draws} object)}
\usage{
get_indices(variables)
}
\description{
Parse the indices from a list of variables (extracted e.g. from a
\code{draws_summary} object or a \code{draws} object)
}
