% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{Generalized_Pareto}
\alias{Generalized_Pareto}
\alias{dgeneralized_pareto}
\alias{pgeneralized_pareto}
\alias{qgpd}
\alias{rgeneralized_pareto}
\title{The Generalized Pareto Distribution}
\usage{
dgeneralized_pareto(x, mu = 0, sigma = 1, k = 0, log = FALSE)

pgeneralized_pareto(
  q,
  mu = 0,
  sigma = 1,
  k = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

qgpd(p, mu = 0, sigma = 1, k = 0, lower.tail = TRUE, log.p = FALSE)

rgeneralized_pareto(n, mu = 0, sigma = 1, k = 0)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scalar location parameter}

\item{sigma}{scalar, positive scale parameter}

\item{k}{scalar shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the generalized Pareto distribution with location equal to \code{mu},
scale equal to \code{sigma}, and shape equal to \code{k}.
}
