% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_constant}
\alias{is_constant}
\title{Check if vector is constant}
\usage{
is_constant(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{(vector) vector to check if is constant}

\item{tol}{(numeric) tolerance to consider two values equal.
Default is \code{.Machine$double.eps}.}
}
\description{
check if a vector is constant, up to a defined tolerance.
}
\keyword{internal}
