% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-.R
\name{is_rvar}
\alias{is_rvar}
\title{Is \code{x} a random variables?}
\usage{
is_rvar(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if \code{x} is an \link{rvar}, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \link{rvar}.
}
\seealso{
\code{\link[=as_rvar]{as_rvar()}} to convert objects to \code{rvar}s.
}
