% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-cast.R
\name{is_rvar_complex}
\alias{is_rvar_complex}
\title{Is \code{x} a complex random variable?}
\usage{
is_rvar_complex(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{\link{rvar}} backed by \code{\link{complex}} draws, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \code{\link{rvar}} backed by \code{\link{complex}} draws.
}
\seealso{
\code{\link[=as_rvar_complex]{as_rvar_complex()}} to convert objects to \code{rvar}s backed by \code{\link{complex}} draws.
}
