% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-factor.R
\name{is_rvar_factor}
\alias{is_rvar_factor}
\alias{is_rvar_ordered}
\title{Is \code{x} a factor random variable?}
\usage{
is_rvar_factor(x)

is_rvar_ordered(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{\link{rvar_factor}} or \code{\link{rvar_ordered}}, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \code{\link{rvar_factor}} or \code{\link{rvar_ordered}}.
}
\seealso{
\code{\link[=as_rvar_factor]{as_rvar_factor()}} and \code{\link[=as_rvar_ordered]{as_rvar_ordered()}} to convert objects to
\code{rvar_factor}s and \code{rvar_ordered}s.
}
