% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-cast.R
\name{is_rvar_integer}
\alias{is_rvar_integer}
\title{Is \code{x} an integer random variable?}
\usage{
is_rvar_integer(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{\link{rvar}} backed by \code{\link{integer}} draws, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \code{\link{rvar}} backed by \code{\link{integer}} draws.
}
\seealso{
\code{\link[=as_rvar_integer]{as_rvar_integer()}} to convert objects to \code{rvar}s backed by \code{\link{integer}} draws.
}
