% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-cast.R
\name{is_rvar_logical}
\alias{is_rvar_logical}
\title{Is \code{x} a logical random variable?}
\usage{
is_rvar_logical(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{\link{rvar}} backed by \code{\link{logical}} draws, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \code{\link{rvar}} backed by \code{\link{logical}} draws.
}
\seealso{
\code{\link[=as_rvar_logical]{as_rvar_logical()}} to convert objects to \code{rvar}s backed by \code{\link{logical}} draws.
}
