% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_mean}
\alias{mcse_mean}
\title{Monte Carlo standard error for the mean}
\usage{
mcse_mean(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute Monte Carlo standard error for mean (expectation) of a
single variable.
}
\examples{
data("draws_eight_schools")
mu <- extract_one_variable_matrix(draws_eight_schools, "mu")
mcse_mean(mu)

}
\references{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
