% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_sd}
\alias{mcse_sd}
\title{Monte Carlo standard error for the standard deviation}
\usage{
mcse_sd(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}
}
\value{
A single numeric value for Monte Carlo standard error
for standard deviation estimate.
}
\description{
Compute Monte Carlo standard error for standard deviation (sd) of a
single parameter using Stirling's approximation and assuming
approximate normality.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
