% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chains.R
\name{merge_chains}
\alias{merge_chains}
\alias{merge_chains.draws_matrix}
\alias{merge_chains.draws_array}
\alias{merge_chains.draws_df}
\alias{merge_chains.draws_list}
\title{Merge Chains of Draws Objects}
\usage{
merge_chains(x, ...)

\method{merge_chains}{draws_matrix}(x, ...)

\method{merge_chains}{draws_array}(x, ...)

\method{merge_chains}{draws_df}(x, ...)

\method{merge_chains}{draws_list}(x, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Merge chains of \code{\link{draws}} objects into a single chain.
}
\examples{
x <- example_draws()
merge_chains(x)

}
