% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variables.R
\name{mutate_variables}
\alias{mutate_variables}
\alias{mutate_variables.draws_matrix}
\alias{mutate_variables.draws_array}
\alias{mutate_variables.draws_df}
\alias{mutate_variables.draws_list}
\title{Mutate Variables in Draws Objects}
\usage{
mutate_variables(.x, ...)

\method{mutate_variables}{draws_matrix}(.x, ...)

\method{mutate_variables}{draws_array}(.x, ...)

\method{mutate_variables}{draws_df}(.x, ...)

\method{mutate_variables}{draws_list}(.x, ...)
}
\arguments{
\item{.x}{A \code{\link{draws}} object.}

\item{...}{Name-value pairs of expressions, each with length 1 or same length
as in the entire input (i.e., number of iterations or draws). The name of
each argument will be the name of a new variable, and the value will be its
corresponding value. Use a \code{NULL} value in \code{mutate_variables} to drop a
variable. New variables overwrite existing variables of the same name.}
}
\value{
Returns a \code{\link{draws}} object of the same format as \code{.x}, with variables mutated
according to the expressions provided in \code{...}.
}
\description{
Mutate variables in a \code{\link{draws}} object.
}
\details{
In order to mutate variables in \code{draws_matrix} and \code{draws_array} objects,
they are transformed to \code{draws_df} objects first and then transformed back
after mutation. As those transformations are quite expensive for larger
number of draws, we recommend using \code{mutate_variables} on \code{draws_df} and
\code{draws_list} objects if speed is an issue.
}
\examples{
x <- as_draws_df(example_draws())
x <- subset(x, variable = c("mu", "tau"))

mutate_variables(x, tau2 = tau^2)
mutate_variables(x, scale = 1.96 * tau, lower = mu - scale)

}
\seealso{
\code{\link{variables}}, \code{\link{rename_variables}}
}
