% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_draws.R
\name{order_draws}
\alias{order_draws}
\title{Order Draws Objects}
\usage{
order_draws(x, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Order \code{draws} objects according to iteration and chain number. By default,
draws objects are ordered but subsetting or extracting parts of them may
leave them in an unordered state.
}
