% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{pareto_k_diagmsg}
\alias{pareto_k_diagmsg}
\title{Pareto-k diagnostic message}
\usage{
pareto_k_diagmsg(diags, ...)
}
\arguments{
\item{diags}{(numeric) named vector of diagnostic values}

\item{...}{unused}
}
\value{
diagnostic message
}
\description{
Given S and scalar and k, form a diagnostic message string
}
