% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{pareto_smooth}
\alias{pareto_smooth}
\alias{pareto_smooth.rvar}
\alias{pareto_smooth.default}
\title{Pareto smoothing}
\usage{
pareto_smooth(x, ...)

\method{pareto_smooth}{rvar}(x, ...)

\method{pareto_smooth}{default}(
  x,
  tail = c("both", "right", "left"),
  r_eff = NULL,
  ndraws_tail = NULL,
  extra_diags = FALSE,
  verbose = FALSE,
  return_smoothed = TRUE,
  ...
)
}
\arguments{
\item{x}{(multiple options) One of:
\itemize{
\item A matrix of draws for a single variable (iterations x chains). See
\code{\link[=extract_variable_matrix]{extract_variable_matrix()}}.
\item An \code{\link{rvar}}.
}}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{tail}{(string) The tail to smooth:
\itemize{
\item \code{"right"}: smooth only the right (upper) tail
\item \code{"left"}: smooth only the left (lower) tail
\item \code{"both"}: smooth both tails and return the maximum k value
}

The default is \code{"both"}.}

\item{r_eff}{(numeric) relative effective sample size estimate. If
\code{r_eff} is omitted, it will be calculated assuming the draws are
from MCMC.}

\item{ndraws_tail}{(numeric) number of draws for the tail. If
\code{ndraws_tail} is not specified, it will be calculated as
max(ceiling(3 * sqrt(length(x))), length(x) / 5).}

\item{extra_diags}{(logical) Should extra Pareto smoothing
diagnostics be included in output? If \code{TRUE}, \code{min_ss},
\code{khat_threshold} and \code{convergence_rate} for the calculated k
value will be returned. Default is \code{FALSE}.}

\item{verbose}{(logical) Should diagnostic messages be printed? If
\code{TRUE}, messages related to Pareto smoothing diagnostics will be
printed. Default is \code{FALSE}.}

\item{return_smoothed}{(logical) Should the smoothed x be
returned? If \code{TRUE}, returns smoothed x. If \code{FALSE}, acts the
same as \code{pareto_khat}. Default is \code{TRUE}.}
}
\value{
List containing vector of smoothed values and vector of
pareto smoothing diagnostics.
}
\description{
Smooth the tail(s) of x by fitting a generalized Pareto
distribution. For further details see Vehtari et al. (2022).
}
\examples{
mu <- extract_variable_matrix(example_draws(), "mu")
pareto_smooth(mu)

d <- as_draws_rvars(example_draws("multi_normal"))
pareto_smooth(d$Sigma)
}
