% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{pit}
\alias{pit}
\alias{pit.default}
\title{Probability integral transfrom}
\usage{
pit(x, ...)

\method{pit}{default}(x, y, loo_weights = NULL, log = TRUE, ...)
}
\arguments{
\item{x}{(draws) A \code{\link{draws_df}} object or one coercible to a \code{draws_df}
object.}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{y}{(observations) A 1D vector of observations. Each element of \code{y}
corresponds to a column in \code{x}.}

\item{loo_weights}{A \code{\link{draws_df}} object or one coercible to a
\code{draws_df} object. \verb{loo_weights`` should have the same dimensions as }x`.}

\item{log}{Are the weights provided in \code{loo_weights} log-weights.}
}
\value{
A numeric vector of length \code{length(y)} containing the PIT values.
}
\description{
The \code{pit()} function computes the empirical probability integral transform
(PIT) of a vector of values with regard to provided draws. If weigths are
provided for the draws, a weighted transformation is computed instead.
}
\details{
The \code{pit()} function computes the probability integral transform of
\code{y} using the empirical cumulative distribution computed from the samples
of \code{x}. These samples can further be provided (log-)weights in

If \code{y} and \code{x} are discrete, randomisation is used to obtain continuous PIT
values. (see, e.g., Czado, C., Gneiting, T., Held, L.: Predictive model
assessment for count data.  Biometrics 65(4), 1254–1261 (2009).)
}
