% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-package.R
\docType{package}
\name{posterior-package}
\alias{posterior-package}
\title{Tools for working posterior distributions}
\description{
Efficiently convert between different representations of posterior samples
and compute modern MCMC diagnostics.
}
