% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.draws_array}
\alias{print.draws_array}
\title{Print \code{draws_array} objects}
\usage{
\method{print}{draws_array}(
  x,
  digits = 2,
  max_iterations = getOption("posterior.max_iterations", 5),
  max_chains = getOption("posterior.max_chains", 8),
  max_variables = getOption("posterior.max_variables", 4),
  ...
)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{digits}{Minimum number of significant digits to print.}

\item{max_iterations}{Maximum number of iterations to print.
Can be controlled globally via the \code{"posterior.max_iterations"} option
(see \code{\link{options}} for details).}

\item{max_chains}{Maximum number of chains to print.
Can be controlled globally via the \code{"posterior.max_chains"} option
(see \code{\link{options}} details).}

\item{max_variables}{Maximum number of variables to print.
Can be controlled globally via the \code{"posterior.max_variables"} option
(see \code{\link{options}} for details).}

\item{...}{Further arguments passed to the underlying \code{\link{print}} methods.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Pretty printing for \code{\link{draws_array}} objects.
}
\examples{
x <- as_draws_array(example_draws())
print(x)

}
