% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.draws_list}
\alias{print.draws_list}
\title{Print \code{draws_list} objects}
\usage{
\method{print}{draws_list}(
  x,
  digits = 2,
  max_iterations = getOption("posterior.max_iterations", 10),
  max_chains = getOption("posterior.max_chains", 2),
  max_variables = getOption("posterior.max_variables", 4),
  reserved = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{draws} object, or another \R object for which the method is defined.}

\item{digits}{Minimum number of significant digits to print.}

\item{max_iterations}{Maximum number of iterations to print.
Can be controlled globally via the \code{"posterior.max_iterations"} option
(see \code{\link{options}} for details).}

\item{max_chains}{Maximum number of chains to print.
Can be controlled globally via the \code{"posterior.max_chains"} option
(see \code{\link{options}} details).}

\item{max_variables}{Maximum number of variables to print.
Can be controlled globally via the \code{"posterior.max_variables"} option
(see \code{\link{options}} for details).}

\item{reserved}{Logical. Indicates whether reserved variables should be
included in the output. Defaults to \code{FALSE}.
See \code{\link{reserved-variables}} for an overview of currently reserved
variable names.}

\item{...}{Further arguments passed to the underlying \code{\link{print}} methods.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Pretty printing for \code{\link{draws_list}} objects.
}
\examples{
x <- as_draws_list(example_draws())
print(x)

}
