% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.draws_matrix}
\alias{print.draws_matrix}
\title{Print \code{draws_matrix} objects}
\usage{
\method{print}{draws_matrix}(
  x,
  digits = 2,
  max_draws = getOption("max_draws", 10),
  max_variables = getOption("max_variables", 8),
  ...
)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{digits}{Minimum number of significant digits to print.}

\item{max_draws}{Maximum number of draws to print.
Can be controlled globally via the \code{"max_draws"} option
(see \code{\link{options}} details).}

\item{max_variables}{Maximum number of variables to print.
Can be controlled globally via the \code{"max_variables"} option
(see \code{\link{options}} for details).}

\item{...}{Further arguments passed to the underlying \code{\link{print}} methods.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Pretty printing for \code{\link{draws_matrix}} objects.
}
\examples{
x <- as_draws_matrix(example_draws())
print(x)

}
