% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\encoding{UTF-8}
\name{print.draws_rvar}
\alias{print.draws_rvar}
\title{Print \code{draws_rvar} objects}
\usage{
\method{print}{draws_rvar}(
  x,
  digits = 2,
  max_variables = getOption("posterior.max_variables", 8),
  summary = getOption("posterior.rvar_summary", "mean_sd"),
  reserved = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{draws} object, or another \R object for which the method is defined.}

\item{digits}{The minimum number of significant digits to print.}

\item{max_variables}{The maximum number of variables to print. Can be
controlled globally via the \code{"posterior.max_variables"}
\link[base:options]{option}.}

\item{summary}{What style of summary to display: \code{"mean_sd"} displays \verb{mean±sd},
\code{"median_mad"} displays \verb{median±mad}. If \code{NULL}, \code{getOption("posterior.rvar_summary")} is
used (default \verb{"mean_sd}).}

\item{reserved}{Should reserved variables be included in the output? Defaults
to \code{FALSE}. See \code{\link{reserved-variables}} for an overview of currently
reserved variable names.}

\item{...}{Further arguments passed to the underlying \code{\link[=print]{print()}} methods.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Pretty printing for \code{\link{draws_rvar}} objects.
}
\examples{
x <- as_draws_rvar(example_draws())
print(x)

}
