% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-print.R
\encoding{UTF-8}
\name{print.rvar}
\alias{print.rvar}
\alias{format.rvar}
\alias{str.rvar}
\title{Print or format a random variable}
\usage{
\method{print}{rvar}(x, ..., summary = NULL, digits = 2, color = TRUE)

\method{format}{rvar}(x, ..., summary = NULL, digits = 2, color = FALSE)

\method{str}{rvar}(object, ..., summary = NULL, vec.len = NULL)
}
\arguments{
\item{x, object}{An \code{\link{rvar}}.}

\item{...}{Further arguments passed to the underlying \code{\link[=print]{print()}} methods.}

\item{summary}{What style of summary to display: \code{"mean_sd"} displays \verb{mean±sd},
\code{"median_mad"} displays \verb{median±mad}. If \code{NULL}, \code{getOption("posterior.rvar_summary")} is
used (default \verb{"mean_sd}).}

\item{digits}{The minimum number of significant digits to print.}

\item{color}{Whether or not to use color when formatting the output. If \code{TRUE},
the \code{\link[pillar:style_subtle]{pillar::style_num()}} functions may be used to produce strings containing
control sequences to produce colored output on the terminal.}

\item{vec.len}{Numeric (>= 0) indicating how many 'first few' elements are
displayed of each vector. If \code{NULL}, defaults to \code{getOption("str")$vec.len},
which defaults to 4.}
}
\value{
For \code{print()} and \code{str()}, an invisible version of the input object.

For \code{format()}, a character vector of the same dimensions as \code{x} where each
entry is of the form \code{"mean±sd"} or \code{"median±mad"}, depending on the value
of \code{summary}.
}
\description{
Printing and formatting methods for \code{\link{rvar}}s.
}
\details{
\code{print()} and \code{str()} print out \code{\link{rvar}} objects by summarizing each element
in the random variable with either its mean±sd or median±mad, depending on
the value of \code{summary}. Both functions use the \code{format()} implementation for
\code{\link{rvar}} objects under the hood, which returns a character vector in the
mean±sd or median±mad form.
}
\examples{

set.seed(5678)
x = rbind(
  cbind(rvar(rnorm(1000, 1)), rvar(rnorm(1000, 2))),
  cbind(rvar(rnorm(1000, 3)), rvar(rnorm(1000, 4)))
)

print(x)
print(x, summary = "median_mad")

str(x)

format(x)

}
