% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar_array-print.R
\name{print.rvar_array}
\alias{print.rvar_array}
\alias{format.rvar_array}
\alias{type_sum.rvar_array}
\alias{pillar_shaft.rvar_array}
\title{Print or format a random variable}
\usage{
\method{print}{rvar_array}(x, ...)

\method{format}{rvar_array}(x, ..., color = FALSE)

\method{type_sum}{rvar_array}(x)

pillar_shaft.rvar_array(x, ...)
}
\arguments{
\item{x}{A list or vector where each entry represents a draw from a distribution}

\item{color}{Whether or not to use color when formatting the output}
}
\value{
An object of class \code{"rvar_array"} representing a random variable.
}
\description{
Printing and formatting methods for random variables of arbitrary objects
}
\details{
The \code{"rvar_array"} class represents random variables of arbitrary objects.
}
