% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_convergence_rate}
\alias{ps_convergence_rate}
\title{Pareto-smoothing convergence rate}
\usage{
ps_convergence_rate(k, S, ...)
}
\arguments{
\item{k}{pareto-k values}

\item{S}{sample size}

\item{...}{unused}
}
\value{
convergence rate
}
\description{
Given S and scalar or array of k's, compute the relative
convergence rate of PSIS estimate RMSE
}
