% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_khat_threshold}
\alias{ps_khat_threshold}
\title{Pareto-smoothing k-hat threshold}
\usage{
ps_khat_threshold(S, ...)
}
\arguments{
\item{S}{sample size}

\item{...}{unused}
}
\value{
threshold
}
\description{
Given sample size S computes khat threshold for reliable Pareto
smoothed estimate (to have small probability of large error). See
section 3.2.4, equation (13).
}
