% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_min_ss}
\alias{ps_min_ss}
\title{Pareto-smoothing minimum sample-size}
\usage{
ps_min_ss(k, ...)
}
\arguments{
\item{k}{pareto k value}

\item{...}{unused}
}
\value{
minimum sample size
}
\description{
Given Pareto-k computes the minimum sample size for reliable Pareto
smoothed estimate (to have small probability of large error)
}
