% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_tail_length}
\alias{ps_tail_length}
\title{Pareto tail length}
\usage{
ps_tail_length(ndraws, r_eff, ...)
}
\arguments{
\item{ndraws}{number of draws}

\item{r_eff}{relative efficiency}

\item{...}{unused}
}
\value{
tail length
}
\description{
Calculate the tail length from number of draws and relative efficiency
r_eff. See Appendix H in Vehtari et al. (2024). This function is
used internally and is exported to be available for other packages.
}
\seealso{
Other helper-functions: 
\code{\link{ps_convergence_rate}()},
\code{\link{ps_khat_threshold}()},
\code{\link{ps_min_ss}()}
}
\concept{helper-functions}
