% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{r_scale}
\alias{r_scale}
\title{Rank values}
\usage{
r_scale(x)
}
\arguments{
\item{x}{(numeric) A scalar, vector, matrix, or array of values.}
}
\value{
A numeric array of ranked values with the same size
and dimension as the input.
}
\description{
Compute ranks for a numeric array, that is, replace each
value by its rank. Average rank for ties are used to conserve the
number of unique values of discrete quantities.
}
\keyword{internal}
