% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variables.R
\name{rename_variables}
\alias{rename_variables}
\alias{rename_variables.draws}
\title{Rename variables in draws objects}
\usage{
rename_variables(.x, ...)

\method{rename_variables}{draws}(.x, ...)
}
\arguments{
\item{.x}{A \link{draws} object.}

\item{...}{One or more expressions separated by commas indicating the variables
to rename. Existing variable names can be unquoted expressions, as in
\code{new_name = old_name}, or character vectors, as in \code{new_name = "old_name"}.}
}
\value{
Returns a \link{draws} object of the same format as \code{.x}, with variables renamed
according to the expressions provided in \code{...}.
}
\description{
Rename variables within a \link{draws} object.
}
\examples{
x <- example_draws()

x
rename_variables(x, mean = mu, sigma = tau)
rename_variables(x, b = `theta[1]`)
rename_variables(x, b = "theta[1]")

}
