% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_draws.R
\name{resample_draws}
\alias{resample_draws}
\alias{resample_draws.draws}
\title{Resample Draws Objects}
\usage{
resample_draws(x, ...)

\method{resample_draws}{draws}(x, weights, method = "stratified", ndraws = NULL, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{weights}{A vector of positive weights of length equal to the number of
draws in \code{draws} (as obtained by method \code{\link{ndraws}}). Weights will be
internally normalized. How exactly the weights are handled depends on the
\code{method} argument.}

\item{method}{Name of the resampling method being applied. Possible choices
are \code{"simple"} for simple random resampling with replacement,
\code{"simple_no_replace"} for simple random resampling without replacement,
\code{"stratified"} for stratified resampling with replacement and
\code{"deterministic"} for deterministic resampling with replacement. Currently,
\code{"stratified"} is the default as it has comparably low variance and bias
with respect to ideal resampling. The latter would sample perfectly
proportional to the weights but is not possible in practice where only a
finite number of draws are available. For more details about resampling
methods, see Kitagawa (1996).}

\item{ndraws}{Number of draws to be returned. By default (\code{NULL}), \code{ndraws}
is set internally to the total number of draws in \code{x} if sensible.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Resample \code{\link{draws}} objects according to given weights obtained,
for instance, through importance sampling.
}
\details{
Upon usage of \code{resample_draws}, chains will be automatically merged
due to subsetting of individual draws (see \code{\link{subset_draws}} for details).
}
\examples{
x <- as_draws_df(example_draws())

# random weights for illustrationary purposes
w <- runif(ndraws(x), 0, 10)

# use default stratified sampling
x_rs <- resample_draws(x, weights = w)
summarise_draws(x_rs, default_summary_measures())

# use simple random sampling
x_rs <- resample_draws(x, weights = w, method = "simple")
summarise_draws(x_rs, default_summary_measures())

}
\references{
Kitagawa, G., Monte Carlo Filter and Smoother for Non-Gaussian Nonlinear '
State Space Models, \emph{Journal of Computational and Graphical Statistics},
5(1):1-25, 1996.
}
