% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved_variables.R
\name{reserved-variables}
\alias{reserved-variables}
\title{Reserved variables}
\description{
Names of reserved variables in the \pkg{posterior} package.
}
\details{
The following variables names are currently reserved for special use cases
in all \code{\link{draws}} formats:
\itemize{
\item \code{.log_weight}: Log weights per draw (see \code{\link{weight_draws}}).
}

Further, specific for the \code{\link{draws_df}} format, there are three additional
reserved variables:
\itemize{
\item \code{.chain}: Chain index per draw
\item \code{.iteration}: Iteration index within each chain
\item \code{.draw}: Draw index across chains
}

More reserved variables may be added in the future.
}
