% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{rhat_rfun}
\alias{rhat_rfun}
\title{Traditional Rhat convergence diagnostic}
\usage{
rhat_rfun(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}
}
\value{
A single numeric value for Rhat.
}
\description{
Compute the Rhat convergence diagnostic for a single parameter
For split-Rhat, call this with split chains.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
