% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_summary.R
\name{rollup_summary}
\alias{rollup_summary}
\alias{rollup_summary.default}
\alias{rollup_summary.draws}
\alias{rollup_summary.data.frame}
\alias{rollup_summary.rollup_summary}
\alias{default_rollups}
\title{Roll up \code{draws_summary} objects by collapsing summaries of non-scalar parameters.}
\usage{
rollup_summary(.x, ...)

\method{rollup_summary}{default}(.x, ...)

\method{rollup_summary}{draws}(.x, ...)

\method{rollup_summary}{data.frame}(.x, ..., variable = NULL, .funs = default_rollups())

\method{rollup_summary}{rollup_summary}(.x, ...)

default_rollups()
}
\arguments{
\item{.x}{(multiple options) The object containing summaries to roll up. One of:
\itemize{
\item a \code{\link{draws_summary}} object such as produced by \code{\link[=summarise_draws]{summarise_draws()}}.
\item a \code{data.frame} with a \code{"variable"} column giving the names of variables,
where all other columns are numeric summaries of those variables.
\item an object with a \code{\link[=summarise_draws]{summarise_draws()}} method, such as a \code{\link{draws}} object,
in which case \code{\link[=summarise_draws]{summarise_draws()}} will be called on \code{.x} and the result
will be rolled up.
\item a \code{\link{rollup_summary}} object such as produced by \code{rollup_summary()}, in
which case variables that have not been rolled up yet may be rolled up.
}}

\item{...}{(multiple options) arguments where the name of each argument is a
summary measure (i.e. column) in \code{.x} and the value is the rollup functions
to apply to that summary measure, specified as one of:
\itemize{
\item bare name of a function
\item a character vector of function names (optionally named).
\item a function formula, as accepted by \code{\link[rlang:as_function]{rlang::as_function()}}.
\item a named list of any of the above.
}

Unnamed arguments in \code{...} specify default rollup functions to apply to all
summary measures that do not have specific rollup functions given in \code{...} or
\code{.funs}.}

\item{variable}{(character vector) base names (without indices) of variables
to roll up. If \code{NULL} (the default), all variables with indices in their names
(e.g. \code{"x[1,2]"}) will be rolled up.}

\item{.funs}{(list) named list where names are summary measures in \code{.x}
and values are the default rollup functions to apply to those summary
measures, unless overridden by \code{...}. As in \code{...}, unnamed elements of this
list give default rollup functions to apply to summary measures that do not
have specific rollup functions given in \code{...} or \code{.funs}.}
}
\value{
A \code{rollup_summary} object, which is a named list of \code{\link{draws_summary}} objects:
\itemize{
\item \code{"unrolled"}: a \code{\link{draws_summary}} of the variables that were not rolled up.
\item \code{"rolled"}: a \code{\link{draws_summary}} of the rolled-up variables. The second
column of this data frame, \code{"dim"}, gives the lengths of the dimensions
of each rolled up variable as a comma-separated character vector. The
remaining columns give the rollups of each summary measure; e.g. if \code{x}
contained a summary measure \code{"mean"} and it was rolled up using the \code{"min"}
and \code{"max"} functions (the default), the output will have a \code{"mean_min"}
and \code{"mean_max"} column.
}
}
\description{
Roll up summaries of draws (e.g. as returned by \code{\link[=summarise_draws]{summarise_draws()}}); that
is, summarise the summaries. By default, summaries of all variables containing
indices (e.g. \code{"x[1]"}) are rolled up, but the \code{variable} parameter can be
used to roll up specific variables only.
}
\details{
If called without specifying additional rollup functions in \code{...},
\code{rollup_summary()} will apply the default rollup functions as determined by
\code{.funs} to the columns in \code{.x} (or, if \code{.x} is not a \code{data.frame}, to the
result of \code{summarise_draws(.x)}).

The default value of \code{.funs} provides several default rollup functions
that will be applied to specific summary measures, unless this is overridden
by entries in \code{...}. For example, \code{ess_bulk} has the default
rollup function \code{"min"} instead of \code{c("min", "max")}, as the minimum
effective sample size is likely of more interest than the maximum.
\code{default_rollups()} gives the complete list of default rollup functions.

Calls to \code{rollup_summary()} can be chained, in which case subsequent
rollups will be applied only to variables that have not already been
rolled up (i.e. the \code{"unrolled"} element; see the description of
\code{rollup_summary} objects below). This makes it possible to provide different
rollup functions for different variables by calling \code{rollup_summary()}
multiple times with different values of the \code{variable} parameter.
}
\examples{
x <- example_draws()

# default summaries show a row for every element in array-like variables
summarise_draws(x)

# you can roll up summaries of array-like variables by rolling up draws
# objects directly; this will apply the default options of summarise_draws()
rollup_summary(x)

# or summarise draws objects first to pick the desired summary measures
# (note that ess_bulk is only rolled up using min by default; see the
# .funs parameter)
ds <- summarise_draws(x, "mean", "sd", "ess_bulk")
rollup_summary(ds)

# rollups work on variables of any dimension
x <- example_draws(example = "multi_normal")
rollup_summary(x)

# you can roll up only some variables
rollup_summary(x, variable = "Sigma")

# you can specify the rollup functions to apply to all summaries by passing
# unnamed parameters ...
rollup_summary(x, "mean", "min")

# ... or use names to specify rollup functions for specific summaries
rollup_summary(x, mean = "sd", median = "min")

# you can pass parameters to rollup functions using anonymous functions
x2 <- draws_rvars(x = c(rvar_rng(rnorm, 5), NA))
rollup_summary(x2, list(min = function(x) min(x, na.rm = TRUE)))

# rollups can be chained to provide different rollup functions to
# different variables
ds <- summarise_draws(x, "mean", "sd")
rs <- rollup_summary(ds, variable = "mu", sd = "min")
rs <- rollup_summary(rs, variable = "Sigma", sd = "max")
rs
}
