% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-dist.R
\name{rvar-functions}
\alias{rvar-functions}
\alias{density.rvar}
\alias{cdf.rvar}
\alias{quantile.rvar}
\title{Density, CDF, and quantile functions of random variables}
\usage{
\method{density}{rvar}(x, at, ...)

\method{cdf}{rvar}(x, q, ...)

\method{quantile}{rvar}(x, probs, ...)
}
\arguments{
\item{x}{(rvar) An \code{\link{rvar}} object.}

\item{...}{Additional arguments passed onto underlying methods:
\itemize{
\item For \code{density()}, these are passed to \code{\link[stats:density]{stats::density()}}.
\item For \code{cdf()}, these are ignored.
\item For \code{quantile()}, these are passed to \code{\link[stats:quantile]{stats::quantile()}}.
}}

\item{q, at}{(numeric vector) One or more quantiles.}

\item{probs}{(numeric vector) One or more probabilities in \verb{[0,1]}.}
}
\value{
A vector of the same length as the input (\code{q}, \code{at}, or \code{probs}) containing
values from the corresponding function of the given \code{\link{rvar}}.
}
\description{
The probability density function (\code{density()}), cumulative distribution
function (\code{cdf()}), and quantile function / inverse CDF (\code{quantile()}) of
an \code{\link{rvar}}.
}
