% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-math.R
\name{rvar-matmult}
\alias{rvar-matmult}
\alias{\%**\%}
\title{Matrix multiplication of random variables}
\usage{
x \%**\% y
}
\arguments{
\item{x}{An \code{\link{rvar}}, \code{\link{numeric}}, or \code{\link{logical}}. Must be 1 or 2-dimensional. If it is a vector,
it is treated as a row vector.}

\item{y}{An \code{\link{rvar}}, \code{\link{numeric}}, or \code{\link{logical}}. Must be 1 or 2-dimensional. If it is a vector,
it is treated as a column vector.}
}
\value{
An \code{\link{rvar}} representing the matrix product of \code{x} and \code{y}.
}
\description{
Matrix multiplication of random variables.
}
\details{
If \code{x} or \code{y} are vectors, they are converted into matrices prior to multiplication, with \code{x}
converted to a row vector and \code{y} to a column vector. Numerics and logicals can be multiplied
by \code{\link{rvar}}s and are broadcasted across all draws of the \code{\link{rvar}} argument. Tensor multiplication
is used to efficiently multiply matrices across draws, so if either \code{x} or \code{y} is an \code{\link{rvar}},
\code{x \%**\% y} will be much faster than \code{rdo(x \%*\% y)}.

Because \code{\link{rvar}} is an S3 class and S3 classes cannot properly override \code{\%*\%}, \code{\link{rvar}}s use
\verb{\%**\%} for matrix multiplication.
}
\examples{

# d has mu (mean vector of length 3) and Sigma (3x3 covariance matrix)
d <- as_draws_rvars(example_draws("multi_normal"))
d$Sigma

# trivial example: multiplication by a non-random matrix
d$Sigma \%**\% diag(1:3)

# Decompose Sigma into R s.t. R'R = Sigma ...
R <- chol(d$Sigma)
# ... and recreate Sigma using matrix multiplication
t(R) \%**\% R

}
