% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-summaries-within-draws.R
\name{rvar-summaries-by-draw}
\alias{rvar-summaries-by-draw}
\alias{Summary.rvar}
\alias{range.rvar}
\alias{rvar_mean}
\alias{rvar_median}
\title{Summaries of random variables over array elements, within draws}
\usage{
\method{Summary}{rvar}(..., na.rm = FALSE)

\method{range}{rvar}(..., na.rm = FALSE)

rvar_mean(x, ...)

rvar_median(x, ...)
}
\arguments{
\item{...}{further arguments passed to underlying functions (e.g., \code{base::mean()}
or \code{base::median()}).}

\item{na.rm}{Should \code{NA} values in the random variable be removed before
computing summaries?}

\item{x}{an \code{\link{rvar}}}
}
\value{
An \code{\link{rvar}} of length 1 (or in the case of \code{range()}, length 2) with the same number
of draws as the input rvar(s) containing the summary statistic computed within
each draw of the input rvar(s).
}
\description{
Compute summaries of random variables within draws, producing a new random variable.
}
\details{
These functions compute statistics within each draw of the random variable.
For summaries over draws (such as expectations), see \link{rvar-summaries}.

Besides \code{rvar_mean()} and \code{rvar_median()}, these standard generics are supported:
\itemize{
\item \code{all()}, \code{any()}
\item \code{sum()}, \code{prod()}
\item \code{min()}, \code{max()}
\item \code{range()}
}
}
\examples{

set.seed(5678)
x = rdo(rnorm(4, mean = 1:4, sd = 2))

# These will give similar results to mean(1:4),
# median(1:4), sum(1:4), prod(1:4), etc
rvar_mean(x)
rvar_median(x)
sum(x)
prod(x)

}
\seealso{
\link{rvar-summaries} for summary functions across draws (e.g. expectations).
\link{rvar-functions} for density, CDF, and quantile functions of random variables.
}
