% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-summaries-over-draws.R
\name{rvar-summaries-over-draws}
\alias{rvar-summaries-over-draws}
\alias{E}
\alias{mean.rvar}
\alias{Pr}
\alias{Pr.default}
\alias{Pr.logical}
\alias{Pr.rvar}
\alias{median.rvar}
\alias{variance.rvar}
\title{Summaries of random variables within array elements, over draws}
\usage{
E(x, ...)

\method{mean}{rvar}(x, ...)

Pr(x, ...)

\method{Pr}{default}(x, ...)

\method{Pr}{logical}(x, ...)

\method{Pr}{rvar}(x, ...)

\method{median}{rvar}(x, ...)

\method{variance}{rvar}(x, ...)
}
\arguments{
\item{x}{an \code{\link{rvar}}}

\item{...}{further arguments passed to underlying functions (e.g., \code{base::mean()}
or \code{base::median()}), such as \code{na.rm}.}
}
\value{
A numeric or logical vector with the same dimensions as the given random variable, where
each entry in the vector is the mean, median, or variance of the corresponding entry in \code{x}.
}
\description{
Compute summaries within elements of an \code{\link{rvar}} and over draws of each element,
producing an array of the same shape as the input random variable (except in
the case of \code{range()}, see 'Details').
}
\details{
Summaries include expectations (\code{E()} or \code{mean()}), probabilities (\code{Pr()}),
medians (\code{median()}), spread (\code{variance()}, \code{sd()}, \code{mad()}), sums and
products (\code{sum()}, \code{prod()}), extrema and ranges (\code{min()}, \code{max()}, \code{range()}),
logical summaries (\code{all()}, \code{any()}), and special value predicates (\code{is.finite()},
\code{is.infinite()}, \code{is.nan()}, \code{is.na()}).

Unless otherwise stated, these functions return a numeric array with the same shape
(same dimensions) as the input \code{\link{rvar}}, \code{x}.

\code{range(x)} return an array with dimensions \code{c(dim(x), 2)}, where the last
dimension contains the minimum and maximum values.

\code{is.infinite(x)}, \code{is.nan(x)}, and \code{is.na(x)} return logical arrays, where each
element is \code{TRUE} if \strong{any} draws in its corresponding element in \code{x} match
the predicate. Each elements in the result of \code{is.finite(x)} is \code{TRUE} if
\strong{all} draws in the corresponding element in \code{x} are finite.

Both \code{E()}, \code{mean()}, and \code{Pr()} return the means of each element in the input.
\code{Pr()} additionally checks that the provided \code{\link{rvar}}
is a logical variable (hence, taking its expectation results in a probability).

For consistency, \code{E()} and \code{Pr()} are also defined for base arrays so that
they can be used as summary functions in \code{summarise_draws()}.
}
\examples{

set.seed(5678)
x = rdo(rnorm(4, mean = 1:4, sd = 2))

# These should all be ~= c(1, 2, 3, 4)
E(x)
mean(x)
median(x)

# This ...
Pr(x < 1.5)
# ... should be about the same as this:
pnorm(1.5, mean = 1:4, sd = 2)

}
\seealso{
\link{rvar-summaries-within-draws} for summary functions within draws.
\link{rvar-functions} for density, CDF, and quantile functions of random variables.

Other rvar-summaries: 
\code{\link{Summary.rvar}()},
\code{\link{rvar_is_finite}()}
}
\concept{rvar-summaries}
