% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar.R
\name{rvar}
\alias{rvar}
\title{Random variables}
\usage{
rvar(x)
}
\arguments{
\item{x}{A list or vector where each entry represents a draw from a distribution}
}
\value{
An object of class \code{"rvar"} representing a random variable.
}
\description{
Random variables of arbitrary objects
}
\details{
The \code{"rvar"} class represents random variables of arbitrary objects.
}
