% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar_array.R
\name{rvar_array}
\alias{rvar_array}
\title{Random variables}
\usage{
rvar_array(x)
}
\arguments{
\item{x}{A vector or array where the last dimension is draws from a distribution}
}
\value{
An object of class \code{"rvar_array"} representing a random variable.
}
\description{
Random variables backed by arrays
}
\details{
The \code{"rvar_array"} class represents random variables as arrays.
}
