% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-variates.R
\name{sd_control_variates}
\alias{sd_control_variates}
\title{Standard deviation estimate with control variates}
\usage{
sd_control_variates(x, gradients = NULL, ...)
}
\arguments{
\item{x}{(rvar) An \code{\link{rvar}}.}

\item{gradients}{the gradients of the natural logarithm of the probability
distribution function for the transformed variables with respect to the
transformed variables, as used for gradient-based methods}
}
\value{
The estimate of \code{sd(x)} using control variates for variance reduction.
}
\description{
Computes the Monte Carlo estimate of the standard deviation using
first order zero-variance control variates (Mira et al, 2013) in the
hopes of obtaining reduced-variance estimates compared to \code{sd()}.
Estimates are based on estimating \code{E[X^2]} and \code{E[x]} separately
with first order zero-variance control variates.
The control variates in this setting are the \code{gradients}.
A sufficient condition for the control variates to have zero expectation
is that the distribution for the transformed variables has tails that decay
faster than polynomially.
}
\examples{
x <- as_draws_df(example_draws())
x$.gradient_mu <- rnorm(ndraws(x))
x$.gradient_sigma <- rnorm(ndraws(x))

grads <- gradients(x)

sd_control_variates(x$mu, gradients = grads)
summarise_draws(x, sd, sd_control_variates, .args = list(gradients = grads))

}
\references{
Antonietta Mira, Reza Solgi and Daniele Imparato (2013).
Zero-variance Markov chain Monte Carlo for Bayesian estimators.
\emph{Statistics and Computing}. 23, 653--662.
doi:10.1007/s11222-012-9344-6
}
