% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{split_chains}
\alias{split_chains}
\title{Split Markov chains in half}
\usage{
split_chains(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A 2D array of draws with split chains.
}
\description{
Split Markov chains in half
}
\keyword{internal}
