% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws_matrix.R
\name{[.draws_matrix}
\alias{[.draws_matrix}
\title{Extract parts of a \code{draws_matrix} object}
\usage{
\method{[}{draws_matrix}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x, i, j, ..., drop}{Same as in the default extraction method but with
\code{drop} being set to \code{FALSE} by default.}
}
\value{
An object of class \code{"draws_matrix"} unless any of the dimensions
was dropped during the extraction.
}
\description{
Extract parts of a \code{draws_matrix} object. They are strictly defined as
matrices (draws x variable) so dropping any of the
dimensions breaks the expected structure of the object. Accordingly, no
dropping of dimensions is done by default even if the extracted slices are of
length 1. If \code{drop} is manually set to \code{TRUE} and any of the dimensions is
actually dropped, this will lead to dropping the \code{"draws_matrix"} class as
well.
}
\keyword{internal}
