% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset-draws}
\alias{subset-draws}
\alias{subset.draws}
\alias{subset.draws_matrix}
\alias{subset.draws_array}
\alias{subset.draws_df}
\alias{subset.draws_list}
\title{Subset Draws Objects}
\usage{
\method{subset}{draws_matrix}(x, variable = NULL, iteration = NULL,
  chain = NULL, draw = NULL, ...)

\method{subset}{draws_array}(x, variable = NULL, iteration = NULL,
  chain = NULL, draw = NULL, ...)

\method{subset}{draws_df}(x, variable = NULL, iteration = NULL,
  chain = NULL, draw = NULL, ...)

\method{subset}{draws_list}(x, variable = NULL, iteration = NULL,
  chain = NULL, draw = NULL, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{variable}{Character vector of variable names to be selected.}

\item{iteration}{Numeric vector of iteration indices to be selected.}

\item{chain}{Numeric vector of chain indices to be selected.}

\item{draw}{Numeric vector of draw indices to be selected.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Subset \code{draws} objects after variables, iteration, chains, and draws
indices.
}
\details{
To ensure that multiple consequtive subsetting operations are working
correctly, \code{subset} repairs \code{draws} objects before and after subsetting
via \code{\link{repair_draws}}.
}
\examples{
x <- example_draws()
subset(x, variable = c("mu", "tau"))
subset(x, chain = 2)
subset(x, iteration = 5:10, chain = 3:4)

}
