% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\alias{thin.draws}
\title{Thin Draws Objects}
\usage{
thin(x, thin, ...)

\method{thin}{draws}(x, thin, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{thin}{A positive integer specifying the period for selecting draws.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Thin \code{draws} objects to reduce their size and autocorrelation of the chains.
}
\examples{
x <- example_draws()
niterations(x)

x <- thin(x, thin = 5)
niterations(x)

}
