% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices.R
\name{variable_indices}
\alias{variable_indices}
\title{Parse the indices from a vector of variables (extracted e.g. from a
\code{draws_summary} object or a \code{draws} object)}
\arguments{
\item{x}{a character vector of variables}
}
\value{
A list of indices for each variable
}
\description{
Parse the indices from a vector of variables (extracted e.g. from a
\code{draws_summary} object or a \code{draws} object)
}
\details{
Assumes that variable indexing is reflected by square brackes in the names
}
