% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-cast.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_ptype2.rvar}
\alias{vec_ptype2.rvar.default}
\alias{vec_cast.rvar}
\alias{vec_cast.rvar.default}
\alias{vec_ptype2.rvar.rvar}
\alias{vec_cast.rvar.rvar}
\alias{vec_ptype2.double.rvar}
\alias{vec_ptype2.rvar.double}
\alias{vec_cast.rvar.double}
\alias{vec_ptype2.integer.rvar}
\alias{vec_ptype2.rvar.integer}
\alias{vec_cast.rvar.integer}
\alias{vec_ptype2.logical.rvar}
\alias{vec_ptype2.rvar.logical}
\alias{vec_cast.rvar.logical}
\title{rvar vctrs compatibility}
\usage{
\method{vec_ptype2}{rvar}(x, y, ...)

\method{vec_ptype2.rvar}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_cast}{rvar}(x, to, ...)

\method{vec_cast.rvar}{default}(x, to, ...)

\method{vec_ptype2.rvar}{rvar}(x, y, ...)

\method{vec_cast.rvar}{rvar}(x, to, ...)

\method{vec_ptype2.double}{rvar}(x, y, ...)

\method{vec_ptype2.rvar}{double}(x, y, ...)

\method{vec_cast.rvar}{double}(x, to, ...)

\method{vec_ptype2.integer}{rvar}(x, y, ...)

\method{vec_ptype2.rvar}{integer}(x, y, ...)

\method{vec_cast.rvar}{integer}(x, to, ...)

\method{vec_ptype2.logical}{rvar}(x, y, ...)

\method{vec_ptype2.rvar}{logical}(x, y, ...)

\method{vec_cast.rvar}{logical}(x, to, ...)
}
\arguments{
\item{x, y, .x}{Vectors}

\item{...}{Further arguments passed to other functions}

\item{x_arg, y_arg}{Argument names for \code{x} and \code{y}}

\item{to}{Type to cast to}

\item{op, .fn}{An operator as a string}
}
\description{
These functions implement compatibility with \link{vctrs-package} for \link{rvar}s.
}
\details{
See the corresponding functions in \link{vctrs-package} for more information.
}
\seealso{
\code{\link[vctrs:vec_cast]{vctrs::vec_cast()}},\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}},\code{\link[vctrs:vec_arith]{vctrs::vec_arith()}},
\code{\link[vctrs:vec_math]{vctrs::vec_math()}}
}
