% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_draws.R
\name{walk_draws}
\alias{walk_draws}
\title{Loop ("walk") over draws}
\usage{
walk_draws(x, expr)
}
\arguments{
\item{x}{(draws) A \code{draws} object or another \R object for which the method
is defined.}

\item{expr}{(expression) A bare expression that can contain references to
variables in \code{x} by name. This expression will be executed once per draw
of \code{x}, where references to variables in \code{x} resolve to the value of that
variable in that draw. The expression supports \link{quasiquotation}.}
}
\description{
Executes an expression once for every draw in a \code{draws} object.
}
\details{
If \code{x} is not in the \code{\link{draws_rvars}} format, it is first converted to that
format. This allows the variables in \code{x} to include their dimensions (i.e,
to act as R vectors and arrays) when being referred to in \code{expr}.

Within \code{expr}, use \code{.draw} to refer to the draw index, which will be a value
between 1 and \code{ndraws(x)}.
}
