% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_draws.R
\name{weights.draws}
\alias{weights.draws}
\title{Extract Weights from Draws Objects}
\usage{
\method{weights}{draws}(object, log = FALSE, normalize = TRUE, ...)
}
\arguments{
\item{object}{A \code{\link{draws}} object.}

\item{log}{Logical. Indicates if weights should be returned on the log scale.
Defaults to \code{FALSE}.}

\item{normalize}{Logical. Indicates if weights should be normalized
in order to sum to 1 on the standard scale. Defaults to \code{TRUE}.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A vector of weights, with one weight per draw.
}
\description{
Extract weights from \code{\link{draws}} objects, with one weight per draw.
See \code{\link{weight_draws}} for details how to add weights to \code{\link{draws}} objects.
}
\examples{
x <- example_draws()

# sample some random weights for illustration
weights <- rexp(ndraws(x))

# add weights
x <- weight_draws(x, weights = weights)

# return normalized weights
weights <- weights(x)
head(weights)

# return normalized log weights
log_weights <- weights(x, log = TRUE)
head(log_weights)

}
\seealso{
\code{\link{weight_draws}}, \code{\link{resample_draws}}
}
