% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-math.R
\name{chol.rvar}
\alias{chol.rvar}
\title{Cholesky decomposition of random matrix}
\usage{
\method{chol}{rvar}(x, ...)
}
\arguments{
\item{x}{(rvar) A 2-dimensional \code{\link{rvar}}.}

\item{...}{Additional parameters passed on to \code{chol.tensor()}}
}
\value{
An \code{\link{rvar}} containing the upper triangular factor of the Cholesky
decomposition, i.e., the matrix \eqn{R} such that \eqn{R'R = x}.
}
\description{
Cholesky decomposition of an \code{\link{rvar}} containing a matrix.
}
