% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{.ess}
\alias{.ess}
\title{Compute the effective sample size}
\usage{
.ess(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute the effective sample size
}
\keyword{internal}
