% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-index.R
\name{draws-index}
\alias{draws-index}
\alias{variables}
\alias{iteration_ids}
\alias{chain_ids}
\alias{draw_ids}
\alias{nvariables}
\alias{niterations}
\alias{nchains}
\alias{ndraws}
\title{Index Draws Objects}
\usage{
variables(x)

iteration_ids(x)

chain_ids(x)

draw_ids(x)

nvariables(x)

niterations(x)

nchains(x)

ndraws(x)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}
}
\description{
Index variables, iterations, chains, and draws.
}
\details{
The methods \code{variables()}, \code{iteration_ids()}, \code{chain_ids()}, and \code{draw_ids()} return
vectors of all variables, iterations, chains, and draws, respectively. In
contrast, the methods \code{nvariables()}, \code{niterations()}, \code{nchains()}, and
\code{ndraws()} return the number of variables, iterations, chains, and draws,
respectively.
}
\examples{
x <- example_draws()

variables(x)
nvariables(x)

iteration_ids(x)
niterations(x)

chain_ids(x)
nchains(x)

draw_ids(x)
ndraws(x)

}
