% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-index.R
\name{draws-index}
\alias{draws-index}
\alias{variables}
\alias{iterations}
\alias{chains}
\alias{draws}
\alias{nvariables}
\alias{niterations}
\alias{nchains}
\alias{ndraws}
\title{Index Draws Objects}
\usage{
variables(x)

iterations(x)

chains(x)

draws(x)

nvariables(x)

niterations(x)

nchains(x)

ndraws(x)
}
\arguments{
\item{x}{An \R object.}
}
\description{
Index variables, iterations, chains, and draws.
}
