% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{draws}
\alias{draws}
\alias{as_draws}
\alias{is_draws}
\title{Transform to \code{draws} objects}
\usage{
as_draws(x, ...)

is_draws(x)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
If possible, a \code{draws} object in the closest supported format to \code{x}.
The formats are linked to in the \strong{See Also} section below.
}
\description{
Try to transform an \R object to a format supported by the \pkg{posterior}
package.
}
\details{
The class \code{"draws"} is the parent class of all supported formats,
which also have their own subclasses of the form \code{"draws_{format}"} (e.g.
\code{"draws_array"}).
}
\seealso{
Other formats: \code{\link{draws_array}},
  \code{\link{draws_df}}, \code{\link{draws_list}},
  \code{\link{draws_matrix}}
}
\concept{formats}
