% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-math.R
\name{E}
\alias{E}
\alias{mean.rvar}
\alias{median.rvar}
\alias{Pr}
\title{Summaries and expectations of random variables}
\usage{
E(x, na.rm = FALSE)

\method{mean}{rvar}(x, ...)

\method{median}{rvar}(x, ...)

Pr(x, na.rm = FALSE)
}
\arguments{
\item{x}{an \code{\link{rvar}}}

\item{na.rm}{Should \code{NA} values in the random variable be removed before
computing summaries?}

\item{...}{further arguments passed to underlying functions (e.g., \code{base::mean()}
or \code{base::median()}).

Both \code{E()}, \code{mean()}, and \code{Pr()} take means over the draws dimension of the provided
random variable. \code{Pr()} additionally checks that the provided \code{\link{rvar}}
is a logical variable (hence, taking its expectation results in a probability).
\code{median()} takes medians.}
}
\value{
A numeric vector with the same dimensions as the given random variable, where
each entry in the vector is the mean or median of the corresponding entry in \code{x}.
}
\description{
Compute expectations (\code{E()} or \code{mean()}), probabilities (\code{Pr()}), and
medians (\code{median()}) from a random variable.
}
\examples{

set.seed(5678)
x = rdo(rnorm(4, mean = 1:4, sd = 2))

# These should all be ~= c(1, 2, 3, 4)
E(x)
mean(x)
median(x)

# This ...
Pr(x < 1.5)
# ... should be about the same as this:
pnorm(1.5, mean = 1:4, sd = 2)

}
\seealso{
\link{Summary.rvar} for summary functions within draws.
\link{rvar-functions} for density, CDF, and quantile functions of random variables.
}
