% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_mean}
\alias{ess_mean}
\title{Effective sample size}
\usage{
ess_mean(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}
}
\value{
A single numeric value for the effective sample size
estimate for mean estimate.
}
\description{
Compute effective sample size estimate for a mean (expectation)
estimate of one parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
