% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_quantile}
\alias{ess_quantile}
\title{Quantile effective sample size}
\usage{
ess_quantile(x, prob)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}

\item{prob}{A single numeric value of probability.}
}
\value{
A single numeric value for the effective sample size for a
quantile estimate corresponding to the probability.
}
\description{
Compute effective sample size estimate for a quantile estimate of
one parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
