% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_sd}
\alias{ess_sd}
\title{Effective sample size for the standard deviation}
\usage{
ess_sd(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}
}
\value{
A single numeric value for the effective sample size
estimate for standard deviation estimate.
}
\description{
Compute an effective sample size estimate for the standard deviation (sd)
estimate of a single parameter. This is defined as minimum of effective
sample size estimate for mean and mean of squared value.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
