% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_sd}
\alias{ess_sd}
\title{Effective sample size for the standard deviation}
\usage{
ess_sd(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute an effective sample size estimate for the standard deviation (SD)
estimate of a single variable. This is defined as minimum of effective
sample size estimate for mean and mean of squared value.
}
\examples{
mu <- extract_one_variable_matrix(example_draws(), "mu")
ess_sd(mu)

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: 
\code{\link{ess_basic}()},
\code{\link{ess_bulk}()},
\code{\link{ess_quantile}()},
\code{\link{ess_tail}()},
\code{\link{mcse_mean}()},
\code{\link{mcse_quantile}()},
\code{\link{mcse_sd}()},
\code{\link{rhat_basic}()},
\code{\link{rhat}()}
}
\concept{diagnostics}
