% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{ess_tail}
\alias{ess_tail}
\title{Tail effective sample size (tail-ESS)}
\usage{
ess_tail(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute tail effective sample size estimate (tail-ESS) for a single variable.
Tail-ESS is useful for generic diagnostic for the sampling
efficiency in the tails of the posterior. It is defined as
the minimum of the effective sample sizes for 5\% and 95\% quantiles.
}
\examples{
mu <- extract_one_variable_matrix(example_draws(), "mu")
ess_tail(mu)

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: 
\code{\link{ess_basic}()},
\code{\link{ess_bulk}()},
\code{\link{ess_quantile}()},
\code{\link{ess_sd}()},
\code{\link{mcse_mean}()},
\code{\link{mcse_quantile}()},
\code{\link{mcse_sd}()},
\code{\link{rhat_basic}()},
\code{\link{rhat}()}
}
\concept{diagnostics}
