% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_draws.R
\name{example_draws}
\alias{example_draws}
\title{Example \code{draws} objects}
\usage{
example_draws()
}
\value{
A \code{draws} object.
}
\description{
Objects for use in examples.
}
\details{
Currently the only example is a \code{\link{draws_array}} object with 100 iterations
from each of 4 Markov chains obtained by fitting the eight schools model
described in Gelman et al. (2013) with \href{https://mc-stan.org}{Stan}. The
variables are:
\itemize{
\item \code{mu}: Overall mean of the eight schools
\item \code{tau}: Standard deviation between schools
\item \code{theta}:  Individual means of each of the eight schools
}
}
\examples{
x <- example_draws()
summarise_draws(x)

}
\references{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
