% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_one_variable_matrix.R
\name{extract_one_variable_matrix}
\alias{extract_one_variable_matrix}
\alias{extract_one_variable_matrix.default}
\alias{extract_one_variable_matrix.draws}
\title{Extract Draws Matrix of one Variable}
\usage{
extract_one_variable_matrix(x, variable, ...)

\method{extract_one_variable_matrix}{default}(x, variable, ...)

\method{extract_one_variable_matrix}{draws}(x, variable, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{variable}{Name of a single variable to extract draws for.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
An \code{matrix} with dimension #iterations x #chains.
}
\description{
Extract a matrix of draws of a single variable with dimension
#iterations x #chains primarily for use in convergence functions
such as \code{\link{rhat}}.
}
\examples{
x <- example_draws()
mu <- extract_one_variable_matrix(x, variable = "mu")
rhat(mu)

}
