% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variable.R, R/extract_variable_matrix.R
\name{extract_variable.draws_rvar}
\alias{extract_variable.draws_rvar}
\alias{extract_variable_matrix}
\alias{extract_variable_matrix.default}
\alias{extract_variable_matrix.draws}
\alias{extract_variable_matrix.draws_rvar}
\title{Extract matrix of a single variable}
\usage{
\method{extract_variable}{draws_rvar}(x, variable, ...)

extract_variable_matrix(x, variable, ...)

\method{extract_variable_matrix}{default}(x, variable, ...)

\method{extract_variable_matrix}{draws}(x, variable, ...)

\method{extract_variable_matrix}{draws_rvar}(x, variable, ...)
}
\arguments{
\item{x}{A \code{draws} object, or another \R object for which the method is defined.}

\item{variable}{Name of a single variable to extract draws for.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
An \code{matrix} with dimension #iterations x #chains.
}
\description{
Extract a matrix of draws of a single variable with dimension
#iterations x #chains, primarily for use in convergence functions
such as \code{\link[=rhat]{rhat()}}.
}
\examples{
x <- example_draws()
mu <- extract_variable_matrix(x, variable = "mu")
rhat(mu)

}
