% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-helpers.R
\name{is_draws}
\alias{is_draws}
\title{Check if argument is a \code{draws} object}
\usage{
is_draws(x)
}
\arguments{
\item{x}{An \R object.}
}
\description{
Check if argument is a \code{draws} object, which constitutes the
parent class of all draws formats supported by \pkg{posterior}.
}
