% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-cast.R
\name{is_rvar}
\alias{is_rvar}
\title{Is \code{x} a random variable?}
\usage{
is_rvar(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{\link{rvar}}, \code{FALSE} otherwise.
}
\description{
Test if \code{x} is an \code{\link{rvar}}.
}
\seealso{
\code{\link[=as_rvar]{as_rvar()}} to convert objects to \code{rvar}s.
}
