% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_mean}
\alias{mcse_mean}
\title{Monte Carlo standard error for the mean}
\usage{
mcse_mean(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}
}
\value{
A single numeric value for Monte Carlo standard error
for mean estimate.
}
\description{
Compute Monte Carlo standard error for mean (expectation) of a
single parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
