% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_quantile}
\alias{mcse_quantile}
\title{Monte Carlo standard error for a quantile}
\usage{
mcse_quantile(x, prob)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (# iter * # chains).}

\item{prob}{A single numeric value of probability.}
}
\value{
A single numeric value for Monte Carlo standard error for a
quantile estimate corresponding to the probability.
}
\description{
Compute Monte Carlo standard error for a quantile estimate of a
single parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
