% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_quantile}
\alias{mcse_quantile}
\alias{mcse_median}
\title{Monte Carlo standard error for quantiles}
\usage{
mcse_quantile(x, probs = c(0.05, 0.95), names = TRUE)

mcse_median(x)
}
\arguments{
\item{x}{One of:
\itemize{
\item A 2D array of draws (iterations x chains) from a single variable.
\item An \code{\link{rvar}}.
}}

\item{probs}{A numeric vector of probabilities.}

\item{names}{Logical. If \code{TRUE}, the result has a names attribute. Set to
\code{FALSE} for speedup with many probs.}
}
\value{
A numeric vector with one element per quantile. If any of the draws is
non-finite, that is, \code{NA}, \code{NaN}, \code{Inf}, or \code{-Inf}, the returned output will
be a vector of (numeric) \code{NA} values. Also, if all draws of a variable are
the same (constant), the returned output will be a vector of (numeric) \code{NA}
values as well. The reason for the latter is that, for constant draws, we
cannot distinguish between variables that are supposed to be constant (e.g.,
a diagonal element of a correlation matrix is always 1) or variables that
just happened to be constant because of a failure of convergence or other
problems in the sampling process.
}
\description{
Compute Monte Carlo standard errors for quantile estimates of a
single variable.
}
\examples{
mu <- extract_variable_matrix(example_draws(), "mu")
mcse_quantile(mu, probs = c(0.1, 0.9))

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: 
\code{\link{ess_basic}()},
\code{\link{ess_bulk}()},
\code{\link{ess_quantile}()},
\code{\link{ess_sd}()},
\code{\link{ess_tail}()},
\code{\link{mcse_mean}()},
\code{\link{mcse_sd}()},
\code{\link{rhat_basic}()},
\code{\link{rhat}()},
\code{\link{rstar}()}
}
\concept{diagnostics}
