% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_quantile}
\alias{mcse_quantile}
\alias{mcse_median}
\title{Monte Carlo standard error for quantiles}
\usage{
mcse_quantile(x, probs, names = TRUE)

mcse_median(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}

\item{probs}{A numeric vector of probabilities.}

\item{names}{Logical. If \code{TRUE}, the result has a names attribute. Set to
\code{FALSE} for speedup with many probs.}
}
\value{
A numeric vector with one element per quantile.
}
\description{
Compute Monte Carlo standard errors for quantile estimates of a
single variable.
}
\examples{
mu <- extract_one_variable_matrix(example_draws(), "mu")
mcse_quantile(mu, probs = c(0.1, 0.9))

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: \code{\link{ess_basic}},
  \code{\link{ess_bulk}}, \code{\link{ess_quantile}},
  \code{\link{ess_sd}}, \code{\link{ess_tail}},
  \code{\link{mcse_mean}}, \code{\link{mcse_sd}},
  \code{\link{rhat_basic}}, \code{\link{rhat}}
}
\concept{diagnostics}
