% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{mcse_sd}
\alias{mcse_sd}
\title{Monte Carlo standard error for the standard deviation}
\usage{
mcse_sd(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute Monte Carlo standard error for standard deviation (SD) of a
single variable using Stirling's approximation and assuming
approximate normality.
}
\examples{
mu <- extract_one_variable_matrix(example_draws(), "mu")
mcse_sd(mu)

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: \code{\link{ess_basic}},
  \code{\link{ess_bulk}}, \code{\link{ess_quantile}},
  \code{\link{ess_sd}}, \code{\link{ess_tail}},
  \code{\link{mcse_mean}}, \code{\link{mcse_quantile}},
  \code{\link{rhat_basic}}, \code{\link{rhat}}
}
\concept{diagnostics}
