% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{pareto_khat}
\alias{pareto_khat}
\alias{pareto_khat.default}
\alias{pareto_khat.rvar}
\title{Pareto khat diagnostic}
\usage{
pareto_khat(x, ...)

\method{pareto_khat}{default}(x, ...)

\method{pareto_khat}{rvar}(x, extra_diags = FALSE, ...)
}
\arguments{
\item{x}{(multiple options) One of:
\itemize{
\item A matrix of draws for a single variable (iterations x chains). See
\code{\link[=extract_variable_matrix]{extract_variable_matrix()}}.
\item An \code{\link{rvar}}.
}}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{extra_diags}{(logical) Should extra Pareto smoothing
diagnostics be included in output? If \code{TRUE}, \code{min_ss},
\code{khat_threshold} and \code{convergence_rate} for the calculated k
value will be returned. Default is \code{FALSE}.}

\item{tail}{(string) The tail to smooth:
\itemize{
\item \code{"right"}: smooth only the right (upper) tail
\item \code{"left"}: smooth only the left (lower) tail
\item \code{"both"}: smooth both tails and return the maximum k value
}

The default is \code{"both"}.}

\item{ndraws_tail}{(numeric) number of draws for the tail. If
\code{ndraws_tail} is not specified, it will be calculated as
ceiling(3 * sqrt(length(x) / r_eff)) if length(x) > 225 and
length(x) / 5 otherwise (see Appendix H in Vehtari et al. (2022)).}

\item{r_eff}{(numeric) relative effective sample size estimate. If
\code{r_eff} is omitted, it will be calculated assuming the draws are
from MCMC.}

\item{verbose}{(logical) Should diagnostic messages be printed? If
\code{TRUE}, messages related to Pareto smoothing diagnostics will be
printed. Default is \code{FALSE}.}
}
\value{
#' List of Pareto smoothing diagnostics:
\itemize{
\item \code{khat}: estimated Pareto k shape parameter, and optionally
\item \code{min_ss}: minimum sample size for reliable Pareto smoothed estimate
\item \code{khat_threshold}: khat-threshold for reliable Pareto smoothed estimate
\item \code{convergence_rate}: Pareto smoothed estimate RMSE convergence rate.
}
}
\description{
Estimate Pareto k value from Pareto smoothing the tail(s) of x. For
further details see Vehtari et al. (2022).
}
\examples{
mu <- extract_variable_matrix(example_draws(), "mu")
pareto_khat(mu)

d <- as_draws_rvars(example_draws("multi_normal"))
pareto_khat(d$Sigma)
}
