% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{pit}
\alias{pit}
\alias{pit.default}
\alias{pit.draws_matrix}
\alias{pit.rvar}
\title{Probability integral transform}
\usage{
pit(x, y, loo_weights = NULL, log = TRUE)

\method{pit}{default}(x, y, loo_weights = NULL, log = TRUE)

\method{pit}{draws_matrix}(x, y, loo_weights = NULL, log = TRUE)

\method{pit}{rvar}(x, y, loo_weights = NULL, log = TRUE)
}
\arguments{
\item{x}{(draws) A \code{\link{draws_matrix}} object or one coercible to a
\code{draws_matrix} object.}

\item{y}{(observations) A 1D vector of observations. Each element of \code{y}
corresponds to a column in \code{x}.}

\item{loo_weights}{A \code{\link{draws_matrix}} object or one coercible to a
\code{draws_matrix} object. \verb{loo_weights`` should have the same dimensions as }x`.}

\item{log}{Are the weights provided in \code{loo_weights} log-weights.}
}
\value{
A numeric vector of length \code{length(y)} containing the PIT values.
}
\description{
The probability integral transform (PIT). LOO-PIT is achieved with weighted sample.
}
\details{
The \code{pit()} function computes the probability integral transform of
\code{y} using the empirical cumulative distribution computed from the samples
of \code{x}. These samples can further be provided (log-)weights in

If \code{y} and \code{x} are discrete, randomisation is used to obtain continuous PIT
values. (see, e.g., Czado, C., Gneiting, T., Held, L.: Predictive model
assessment for count data.  Biometrics 65(4), 1254–1261 (2009).)
}
