% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-package.R
\docType{package}
\name{posterior-package}
\alias{posterior-package}
\alias{posterior}
\title{Tools for working with posterior (and prior) distributions}
\description{
\if{html}{
   \figure{stanlogo.png}{options: width="50"}
   \url{https://mc-stan.org/posterior}
}

The \pkg{posterior} package is intended to provide useful tools
for both users and developers of packages for fitting Bayesian models or
working with output from Bayesian models. The primary goals of the package
are to:
\itemize{
\item Efficiently convert between many different useful formats of
draws (samples) from posterior or prior distributions.
\item Provide consistent methods for operations commonly performed on draws,
for example, subsetting, binding, or mutating draws.
\item Provide various summaries of draws in convenient formats.
\item Provide lightweight implementations of state of the art posterior inference
diagnostics.
}
}
\section{Package options}{


The following options are used to format and print \link{draws} objects,
as in \code{print.draws_array()}, \code{print.draws_df()}, \code{print.draws_list()},
and \code{print.draws_matrix()}:
\itemize{
\item \code{posterior.max_draws}: Maximum number of draws to print.
\item \code{posterior.max_iterations}: Maximum number of iterations to print.
\item \code{posterior.max_chains}: Maximum number of chains to print.
\item \code{posterior.max_variables}: Maximum number of variables to print.
}
}

