% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-print.R
\name{print.rvar}
\alias{print.rvar}
\alias{format.rvar}
\alias{type_sum.rvar}
\alias{pillar_shaft.rvar}
\title{Print or format a random variable}
\usage{
\method{print}{rvar}(x, ...)

\method{format}{rvar}(x, ..., color = FALSE)

\method{type_sum}{rvar}(x)

pillar_shaft.rvar(x, ...)
}
\arguments{
\item{x}{A list or vector where each entry represents a draw from a distribution}

\item{color}{Whether or not to use color when formatting the output}
}
\value{
An object of class \code{"rvar"} representing a random variable.
}
\description{
Printing and formatting methods for random variables of arbitrary objects
}
\details{
The \code{"rvar"} class represents random variables of arbitrary objects.
}
