% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_convergence_rate}
\alias{ps_convergence_rate}
\title{Pareto convergence rate}
\usage{
ps_convergence_rate(k, ndraws, ...)
}
\arguments{
\item{k}{pareto-k values}

\item{ndraws}{number of draws}

\item{...}{unused}
}
\value{
convergence rate
}
\description{
Given number of draws and scalar or array of k's, compute the
relative convergence rate of PSIS estimate RMSE. See Appendix B of
Vehtari et al. (2024). This function is exported to be usable by
other packages. For user-facing diagnostic functions, see
\code{\link{pareto_convergence_rate}} and \code{\link{pareto_diags}}.
}
\seealso{
Other helper-functions: 
\code{\link{ps_khat_threshold}()},
\code{\link{ps_min_ss}()},
\code{\link{ps_tail_length}()}
}
\concept{helper-functions}
