% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_khat_threshold}
\alias{ps_khat_threshold}
\title{Pareto k-hat threshold}
\usage{
ps_khat_threshold(ndraws, ...)
}
\arguments{
\item{ndraws}{number of draws}

\item{...}{unused}
}
\value{
threshold
}
\description{
Given number of draws, computes khat threshold for reliable Pareto
smoothed estimate (to have small probability of large error). See
section 3.2.4, equation (13) of Vehtari et al. (2024). This
function is exported to be usable by other packages. For
user-facing diagnostic functions, see \code{\link{pareto_khat_threshold}} and
\code{\link{pareto_diags}}.
}
\seealso{
Other helper-functions: 
\code{\link{ps_convergence_rate}()},
\code{\link{ps_min_ss}()},
\code{\link{ps_tail_length}()}
}
\concept{helper-functions}
