% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_min_ss}
\alias{ps_min_ss}
\title{Pareto-smoothing minimum sample-size}
\usage{
ps_min_ss(k, ...)
}
\arguments{
\item{k}{pareto k value}

\item{...}{unused}
}
\value{
minimum sample size
}
\description{
Given Pareto-k computes the minimum sample size for reliable Pareto
smoothed estimate (to have small probability of large error). See
section 3.2.3 in Vehtari et al. (2024). This function is exported
to be usable by other packages. For user-facing diagnostic functions, see
\code{\link{pareto_min_ss}} and \code{\link{pareto_diags}}.
}
\seealso{
Other helper-functions: 
\code{\link{ps_convergence_rate}()},
\code{\link{ps_khat_threshold}()},
\code{\link{ps_tail_length}()}
}
\concept{helper-functions}
