% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smooth.R
\name{ps_tail_sorted_log_ratios}
\alias{ps_tail_sorted_log_ratios}
\title{Pareto smooth sorted tail of log ratios}
\usage{
ps_tail_sorted_log_ratios(x, cutoff, smooth_draws = TRUE, ...)
}
\arguments{
\item{x}{Vector of tail elements already sorted in ascending order. Already shifted to account for overflow.}

\item{cutoff}{Largest value smaller than the tail values.}

\item{smooth_draws}{(logical) Should the tails be smoothed? Default is
\code{TRUE}. If \code{FALSE}, \code{k} will be calculated but \code{x} will remain untouched.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Function to Pareto smooth a sorted tail of log ratios. Exported
for usage in other packages, not by users.
}
\seealso{
\code{\link{ps_tail}} for the general tail smoothing function.
}
