% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{quantile2}
\alias{quantile2}
\title{Compute Quantiles}
\usage{
quantile2(x, probs = c(0.05, 0.95), names = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains) for a single variable.}

\item{probs}{A numeric vector of probabilities.}

\item{names}{Logical. If \code{TRUE}, the result has a names attribute. Set to
\code{FALSE} for speedup with many probs.}

\item{na.rm}{Logical. If \code{TRUE}, any \code{NA} and \code{NaN}'s are removed from \code{x}
before the quantiles are computed.}

\item{...}{Further arguments passed to \code{\link[=quantile]{quantile()}}.}
}
\description{
Compute quantiles of a sample and return them in a format consistent
with other summary functions in the \pkg{posterior} package.
}
\examples{
mu <- extract_variable_matrix(example_draws(), "mu")
quantile2(mu)

}
