% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-dist.R, R/rvar-summaries-over-draws.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cdf}
\alias{variance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distributional}{\code{\link[distributional]{cdf}}, \code{\link[distributional]{variance}}}
}}

