% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_draws.R
\name{repair_draws}
\alias{repair_draws}
\title{Repair Indices of Draws Objects}
\usage{
repair_draws(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Repair indices of draws objects so that iterations, chains, and draws
are continuously and consistently numbered.
}
