% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{rhat}
\alias{rhat}
\title{Rhat convergence diagnostic}
\usage{
rhat(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute Rhat convergence diagnostic as the maximum of rank normalized
split-Rhat and rank normalized folded-split-Rhat for a single variable
as proposed in Vehtari et al. (2019).
}
\examples{
data("draws_eight_schools")
mu <- extract_one_variable_matrix(draws_eight_schools, "mu")
rhat(mu)

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
