% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{Rhat}
\alias{Rhat}
\title{Rhat convergence diagnostic}
\usage{
Rhat(x)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}
}
\value{
A single numeric value for the effective sample size.
}
\description{
Compute Rhat convergence diagnostic as the maximum of rank normalized
split-Rhat and rank normalized folded-split-Rhat for a single parameter
as proposed in Vehtari et al. (2019).
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
