% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{rhat}
\alias{rhat}
\title{Rhat convergence diagnostic}
\usage{
rhat(x)
}
\arguments{
\item{x}{A 2D array of draws (iterations x chains).}
}
\value{
A single numeric value.
}
\description{
Compute Rhat convergence diagnostic as the maximum of rank normalized
split-Rhat and rank normalized folded-split-Rhat for a single variable
as proposed in Vehtari et al. (2019).
}
\examples{
mu <- extract_variable_matrix(example_draws(), "mu")
rhat(mu)

}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
\seealso{
Other diagnostics: 
\code{\link{ess_basic}()},
\code{\link{ess_bulk}()},
\code{\link{ess_quantile}()},
\code{\link{ess_sd}()},
\code{\link{ess_tail}()},
\code{\link{mcse_mean}()},
\code{\link{mcse_quantile}()},
\code{\link{mcse_sd}()},
\code{\link{rhat_basic}()}
}
\concept{diagnostics}
