% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{Rhat_basic}
\alias{Rhat_basic}
\title{Basic version of the Rhat convergence diagnostic}
\usage{
Rhat_basic(x, split = TRUE)
}
\arguments{
\item{x}{A 2D array \emph{without} warmup draws (#iterations * #chains).}

\item{split}{Logical. If \code{TRUE}, compute Rhat on split chains.}
}
\value{
A single numeric value for Rhat.
}
\description{
Compute the basic Rhat convergence diagnostic for a single parameter as
described in Gelman et al. (2013). For practical applications, we strongly
recommend the improved Rhat convergence diagnostic implemented in
\code{\link{Rhat}}.
}
\references{
Andrew Gelman, John B. Carlin, Hal S. Stern, David B. Dunson, Aki Vehtari and
Donald B. Rubin (2013). Bayesian Data Analysis, Third Edition. Chapman and
Hall/CRC.
}
