% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-math.R
\name{rvar-matmult}
\alias{rvar-matmult}
\alias{\%**\%}
\title{Matrix multiplication of random variables}
\usage{
x \%**\% y
}
\arguments{
\item{x}{An \link{rvar}, \link{numeric}, or \link{logical}. Must be 1 or 2-dimensional. If it is a vector,
it is treated as a row vector.}

\item{y}{An \link{rvar}, \link{numeric}, or \link{logical}. Must be 1 or 2-dimensional. If it is a vector,
it is treated as a column vector.}
}
\value{
An \link{rvar} representing the matrix product of \code{x} and \code{y}.
}
\description{
Matrix multiplication of random variables.
}
\details{
If \code{x} or \code{y} are vectors, they are converted into matrices prior to multiplication, with \code{x}
converted to a row vector and \code{y} to a column vector. Numerics and logicals can be multiplied
by \link{rvar}s and are broadcasted across all draws of the \link{rvar} argument. Tensor multiplication
is used to efficiently multiply matrices across draws, so if either \code{x} or \code{y} is an \link{rvar},
\code{x \%**\% y} will be much faster than \code{rdo(x \%*\% y)}.
}
