% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvar-summaries-over-draws.R
\name{rvar-summaries-over-draws}
\alias{rvar-summaries-over-draws}
\alias{E}
\alias{mean.rvar}
\alias{Pr}
\alias{Pr.default}
\alias{Pr.logical}
\alias{Pr.rvar}
\alias{median.rvar}
\alias{variance.rvar}
\title{Summaries of random variables within array elements, over draws}
\usage{
E(x, ...)

\method{mean}{rvar}(x, ...)

Pr(x, ...)

\method{Pr}{default}(x, ...)

\method{Pr}{logical}(x, ...)

\method{Pr}{rvar}(x, ...)

\method{median}{rvar}(x, ...)

\method{variance}{rvar}(x, ...)
}
\arguments{
\item{x}{an \code{\link{rvar}}}

\item{...}{further arguments passed to underlying functions (e.g., \code{base::mean()}
or \code{base::median()}), such as \code{na.rm}.}
}
\value{
A numeric vector with the same dimensions as the given random variable, where
each entry in the vector is the mean, median, or variance of the corresponding entry in \code{x}.
}
\description{
Compute expectations (\code{E()} or \code{mean()}), probabilities (\code{Pr()}),
medians (\code{median()}), and variances (\code{variance()}) from a random variable.
}
\details{
Both \code{E()}, \code{mean()}, and \code{Pr()} take means over the draws dimension of the provided
random variable. \code{Pr()} additionally checks that the provided \code{\link{rvar}}
is a logical variable (hence, taking its expectation results in a probability).
\code{median()} takes medians, and \code{variance()} takes variances.

For consistency, \code{E()} and \code{Pr()} are also defined for base arrays so that
they can be used as summary functions in \code{summarise_draws()}.
}
\examples{

set.seed(5678)
x = rdo(rnorm(4, mean = 1:4, sd = 2))

# These should all be ~= c(1, 2, 3, 4)
E(x)
mean(x)
median(x)

# This ...
Pr(x < 1.5)
# ... should be about the same as this:
pnorm(1.5, mean = 1:4, sd = 2)

}
\seealso{
\link{rvar-summaries-within-draws} for summary functions within draws.
\link{rvar-functions} for density, CDF, and quantile functions of random variables.

Other rvar-summaries: 
\code{\link{Summary.rvar}()},
\code{\link{rvar_is_finite}()}
}
\concept{rvar-summaries}
