% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chains.R
\name{split_chains}
\alias{split_chains}
\title{Split Chains}
\usage{
split_chains(x, ...)
}
\arguments{
\item{x}{(draws) A \code{draws} object or another \R object for which the method
is defined.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Split chains by halving the number of iterations per chain and doubling the
number of chains.
}
\examples{
x <- example_draws()
niterations(x)
nchains(x)

x <- split_chains(x)
niterations(x)
nchains(x)

}
