% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_draws.R
\name{subset_draws}
\alias{subset_draws}
\alias{subset_draws.draws_matrix}
\alias{subset_draws.draws_array}
\alias{subset_draws.draws_df}
\alias{subset_draws.draws_list}
\alias{subset.draws}
\title{Subset \code{draws} objects}
\usage{
subset_draws(x, ...)

\method{subset_draws}{draws_matrix}(
  x,
  variable = NULL,
  iteration = NULL,
  chain = NULL,
  draw = NULL,
  regex = FALSE,
  unique = TRUE,
  ...
)

\method{subset_draws}{draws_array}(
  x,
  variable = NULL,
  iteration = NULL,
  chain = NULL,
  draw = NULL,
  regex = FALSE,
  unique = TRUE,
  ...
)

\method{subset_draws}{draws_df}(
  x,
  variable = NULL,
  iteration = NULL,
  chain = NULL,
  draw = NULL,
  regex = FALSE,
  unique = TRUE,
  ...
)

\method{subset_draws}{draws_list}(
  x,
  variable = NULL,
  iteration = NULL,
  chain = NULL,
  draw = NULL,
  regex = FALSE,
  unique = TRUE,
  ...
)

\method{subset}{draws}(x, ...)
}
\arguments{
\item{x}{An \R object for which the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}

\item{variable}{Character vector of variable names to be selected.
All elements of non-scalar variables can be selected at once.}

\item{iteration}{Numeric vector of iteration indices to be selected.}

\item{chain}{Numeric vector of chain indices to be selected.}

\item{draw}{Numeric vector of draw indices to be selected.
Subsetting draw indices will lead to an automatic merging of chains
via \code{\link{merge_chains}}.}

\item{regex}{Logical. Indicates whether \code{variable} should be treated as a
(vector of) regular expressions. Any variable in \code{x} matching at least one
of the regular expressions will be selected. Defaults to \code{FALSE}.}

\item{unique}{Logical. Indicates whether duplicated selection of chains,
iterations, or draws is possible. If \code{TRUE} (the default) only
unique chains, iterations, and draws are selected regardless of how
often they appear in the respective selecting arguments.}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Subset \code{\link{draws}} objects by variables, iterations, chains, and draws indices.
}
\details{
To ensure that multiple consecutive subsetting operations work correctly,
\code{subset()} \emph{\link[=repair_draws]{repairs}} the \code{draws} object before and after
subsetting.
}
\examples{
x <- example_draws()
subset_draws(x, variable = c("mu", "tau"))
subset_draws(x, chain = 2)
subset_draws(x, iteration = 5:10, chain = 3:4)

# extract the first chain twice
subset_draws(x, chain = c(1, 1), unique = FALSE)

# extract all elements of 'theta'
subset_draws(x, variable = "theta")

}
