% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_draws.R
\name{thin_draws}
\alias{thin_draws}
\alias{thin}
\alias{thin_draws.draws}
\alias{thin_draws.rvar}
\title{Thin \code{draws} objects}
\usage{
thin_draws(x, thin, ...)

\method{thin_draws}{draws}(x, thin, ...)

\method{thin_draws}{rvar}(x, thin, ...)
}
\arguments{
\item{x}{(draws) A \code{draws} object or another \R object for which the method
is defined.}

\item{thin}{(positive integer) The period for selecting draws.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A \code{draws} object of the same class as \code{x}.
}
\description{
Thin \code{\link{draws}} objects to reduce their size and autocorrelation in the chains.
}
\examples{
x <- example_draws()
niterations(x)

x <- thin_draws(x, thin = 5)
niterations(x)

}
