% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{z_scale}
\alias{z_scale}
\title{Rank normalization}
\usage{
z_scale(x, c = 3/8)
}
\arguments{
\item{x}{(numeric) A scalar, vector, matrix, or array of values.}

\item{c}{(numeric) Fractional offset used in the back-transformation of ranks.
Defaults to \code{3/8}.}
}
\value{
A numeric array of rank normalized values with the same size
and dimension as the input.
}
\description{
Compute rank normalization for a numeric array. First replace each
value by its rank. Average rank for ties are used to conserve the
number of unique values of discrete quantities. Second, normalize
ranks via the inverse normal transformation.
}
\keyword{internal}
